% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STModAsreml.R
\name{bestSpatMod}
\alias{bestSpatMod}
\title{Helper function for calculating best spatial model using asreml.}
\usage{
bestSpatMod(TD, traits, what = c("fixed", "random"), regular = TRUE,
  criterion = "AIC", useCheckId = FALSE, design = "rowcol",
  covariates = NULL, fixedForm, randomForm, ...)
}
\description{
Helper function for calculating best spatial model using asreml.
}
\keyword{internal}
