% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTD.R
\name{checkTDMeta}
\alias{checkTDMeta}
\title{Helper function for checking metadata structure for TD objects.}
\usage{
checkTDMeta(trLocation = NULL, trDate = NULL, trDesign = NULL,
  trLat = NULL, trLong = NULL, trPlWidth = NULL, trPlLength = NULL)
}
\description{
Helper function for checking metadata structure for TD objects.
}
\keyword{internal}
