% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chkLastIter}
\alias{chkLastIter}
\title{Helper function for checking whether error message about 1% change on
last iteration for asreml is worth mentioning as a warning.
If the corresponding parameter is close to zero and then changes of 1%
or more can be expected and are ok.}
\usage{
chkLastIter(model)
}
\description{
Helper function for checking whether error message about 1% change on
last iteration for asreml is worth mentioning as a warning.
If the corresponding parameter is close to zero and then changes of 1%
or more can be expected and are ok.
}
\keyword{internal}
