% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeMegaEnv.R
\name{combCor}
\alias{combCor}
\title{Helper function for computing combined correlations for a pair of variables.
Following the notation of Charter (1993).}
\usage{
combCor(Xi, Yi, SXi, SYi, ni, ri)
}
\description{
Helper function for computing combined correlations for a pair of variables.
Following the notation of Charter (1993).
}
\keyword{internal}
