% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeMegaEnv.R
\name{combLocs}
\alias{combLocs}
\title{Helper function for computing the combined correlation for 2 locations.}
\usage{
combLocs(l1, l2, ammi, r0, Xi, SXi)
}
\description{
Helper function for computing the combined correlation for 2 locations.
}
\keyword{internal}
