% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STExtract.R
\name{createBaseData}
\alias{createBaseData}
\title{Helper function for creating baseData}
\usage{
createBaseData(TD, predicted, keep = NULL, useRepId = FALSE,
  bdPred = FALSE)
}
\description{
Helper function for creating baseData
}
\keyword{internal}
