% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createReport}
\alias{createReport}
\title{Helper function for creating the actual report}
\usage{
createReport(x, reportName, outfile, ...)
}
\description{
Helper function for creating the actual report
}
\keyword{internal}
