% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarComp.R
\name{createVarComp}
\alias{createVarComp}
\title{S3 class varComp}
\usage{
createVarComp(
  fitMod,
  modDat,
  trait,
  nestingFactor,
  useLocYear,
  useRegionLocYear,
  fullRandVC,
  aovFullFixedMod,
  engine,
  confoundVars,
  diagTabs
)
}
\arguments{
\item{fitMod}{A fitted variance components model.}

\item{modDat}{A data.frame containing the data used in fitting the model.}
}
\description{
Function for creating objects of S3 class varComp.\cr
\code{\link{print}}, \code{\link{summary}}, and \code{\link{plot}} methods
are available.
}
\keyword{internal}
