% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarComp.R
\name{diagnostics}
\alias{diagnostics}
\title{Get diagnostics for an object of class varComp}
\usage{
diagnostics(varComp)
}
\arguments{
\item{varComp}{An object of class varComp.}
}
\value{
A list of tables is invisibly returned.
}
\description{
Get the diagnostics for the model fitted. This will print a table of
combinations missing in the data. For each random factor in the model a
table is printed.
}
\examples{
## Fit a mixed model.
geVarComp <- gxeVarComp(TD = TDMaize, trait = "yld")

## Display diagnostics.
diagnostics(geVarComp)

}
\seealso{
Other Mixed model analysis: 
\code{\link{CRDR}()},
\code{\link{correlations}()},
\code{\link{gxeVarComp}()},
\code{\link{herit}()},
\code{\link{plot.varComp}()},
\code{\link{predict.varComp}()},
\code{\link{vc}()}
}
\concept{Mixed model analysis}
