% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escapeLatex}
\alias{escapeLatex}
\title{Function for escaping special LaTeX characters}
\usage{
escapeLatex(x, newlines = FALSE, spaces = FALSE)
}
\description{
Taken from knitr. Copied since it is an internal knitr function.
}
\keyword{internal}
