% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractAsreml}
\alias{extractAsreml}
\title{Extract statistics from model fitted using asreml}
\usage{
extractAsreml(SSA, traits = SSA$traits, what = "all", keep = NULL,
  useRepId, restore = FALSE)
}
\description{
Extract statistics from model fitted using asreml
}
\keyword{internal}
