% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STExtract.R
\name{extractLme4}
\alias{extractLme4}
\title{Extract statistics from model fitted using lme4}
\usage{
extractLme4(SSA, traits = SSA$traits, what = "all", keep = NULL,
  useRepId, restore = FALSE)
}
\description{
Extract statistics from model fitted using lme4
}
\keyword{internal}
