% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STExtract.R
\name{extractSommer}
\alias{extractSommer}
\title{Extract statistics from model fitted using sommer}
\usage{
extractSommer(SSA, traits = SSA$traits, what = "all", keep = NULL,
  useRepId)
}
\description{
Extract statistics from model fitted using sommer
}
\keyword{internal}
