% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STExtract.R
\name{extractSpATS}
\alias{extractSpATS}
\title{Extract statistics from model fitted using SpATS}
\usage{
extractSpATS(SSA, traits = SSA$traits, what = "all", keep = NULL,
  useRepId, restore = FALSE)
}
\description{
Extract statistics from model fitted using SpATS
}
\keyword{internal}
