% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractVarComp}
\alias{extractVarComp}
\title{Function for extracting the table with variance components from a model in
a nicely printable format.}
\usage{
extractVarComp(model, engine)
}
\description{
Function for extracting the table with variance components from a model in
a nicely printable format.
}
\keyword{internal}
