% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFW.R
\name{fitted.FW}
\alias{fitted.FW}
\title{Extract fitted values.}
\usage{
\method{fitted}{FW}(object, ...)
}
\arguments{
\item{object}{An object of class FW}

\item{...}{Not used.}
}
\value{
A data.frame with fitted values.
}
\description{
Extract the fitted values for a fitted Finlay-Wilkinson model.
}
\examples{
## Run Finlay-Wilkinson analysis.
geFW <- gxeFw(TD = TDMaize, trait = "yld")

## Extract fitted values.
fitFW <- fitted(geFW)
head(fitFW)

}
\seealso{
Other Finlay-Wilkinson: 
\code{\link{gxeFw}()},
\code{\link{plot.FW}()},
\code{\link{report.FW}()},
\code{\link{residuals.FW}()}
}
\concept{Finlay-Wilkinson}
