% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarCov.R
\name{fitted.varCov}
\alias{fitted.varCov}
\title{Extract fitted values.}
\usage{
\method{fitted}{varCov}(object, ...)
}
\arguments{
\item{object}{An object of class varCov}

\item{...}{Not used.}
}
\description{
Extract the fitted values for an object of class varCov.
}
