% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTD.R
\name{getMeta}
\alias{getMeta}
\title{Extract metadata from TD objects}
\usage{
getMeta(TD)
}
\arguments{
\item{TD}{An object of class TD.}
}
\value{
A data.frame containing the metadata for all trials in TD.
}
\description{
Function for extracting metadata as a data.frame from objects of class TD.
Location, data, design, latitude, longitude, plotWidth and plotLength for
all trials in TD will be extracted and returned as a data.frame.
}
