% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeAmmi.R
\name{gxeAmmi}
\alias{gxeAmmi}
\title{AMMI analysis}
\usage{
gxeAmmi(TD, trait, nPC = 2, center = TRUE, scale = FALSE)
}
\arguments{
\item{TD}{an object of class \code{\link{TD}}.}

\item{trait}{A character string specifying the trait to be analyzed.}

\item{nPC}{An integer specifying the number of principal components used
as multiplicative term of genotype-by-environment interaction.}

\item{center}{Should the variables be shifted to be zero centered?}

\item{scale}{Should the variables be scaled to have unit variance?}
}
\value{
An object of class \code{\link{AMMI}}, a list containing:
\item{envScores}{A matrix with environmental scores.}
\item{genoScores}{A matrix with genotypic scores.}
\item{importance}{A data.frame containing the importance of the principal
components.}
\item{anova}{A data.frame containing anova scores of the AMMI analysis.}
\item{fitted}{A matrix containing fitted values from the AMMI model.}
\item{trait}{A character string containing the analyzed trait.}
\item{envMean}{A numerical vector containing the environmental means.}
\item{genoMean}{A numerical vector containing the genotypic means.}
\item{overallMean}{A numerical value containing the overall mean.}
}
\description{
This function fits a model which involves the Additive Main effects (i.e.
genotype and environment) along with the Multiplicative Interaction effects
of principal component analysis (PCA).
}
\examples{
## Run AMMI analysis on TDMaize.
geAmmi <- gxeAmmi(TD = TDMaize, trait = "yld")
## Summarize results.
summary(geAmmi)
## Create a biplot of genotypes and environment interaction with PC1 and PC2.
plot(geAmmi, plotType = "AMMI2")
## Create a pdf report summarizing the results.
\dontrun{
report(geAmmi, outfile = "./testReports/reportAmmi.pdf")
}

}
\references{
Gauch H.G. (1992) Statistical Analysis of Regional Yield Trials:
AMMI Analysis of Factorial Designs. Elsevier, Amsterdam.
}
\seealso{
\code{\link{AMMI}}, \code{\link{plot.AMMI}},
\code{\link{report.AMMI}}
}
