% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeMegaEnv.R
\name{gxeMegaEnv}
\alias{gxeMegaEnv}
\title{Form mega-environments based on winning genotypes from an AMMI model}
\usage{
gxeMegaEnv(TD, trials = names(TD), trait, useWinGeno = TRUE,
  method = c("max", "min"), cutOff = 0.8, sumTab = TRUE)
}
\arguments{
\item{TD}{An object of class \code{\link{TD}}.}

\item{trials}{A character string specifying the trials to be analyzed. If
not supplied all trials are used in the analysis.}

\item{trait}{A character string specifying the trait to be analyzed.}

\item{useWinGeno}{Should only the best genotype per trail be used for
determining mega environments? If \code{FALSE} clustering of trials is used.}

\item{method}{A character string indicating the criterion to determine
the best genotype per environment, either \code{"max"} or \code{"min"}.}

\item{cutOff}{A numerical value indicating the proportion of best genotypes
per locationh used in the calculation. I.e. a value of 0.8 indicates that
the best 80\% genotypes will be used.}

\item{sumTab}{Should a summary table be printed?}
}
\value{
The input object of class \code{\link{TD}} with an added extra
column megaEnv.
}
\description{
This function fits an AMMI model and then using the fitted values produces
a new factor based on the winning genotype in each environment. This factor
is added as a column megaEnv to the input data. If a column megaEnv already
exists the existing data is overwritten with a warning.
}
\examples{
## Calculate mega-environments for TDMaize and print a summary of the results.
TDmegaEnv <- gxeMegaEnv(TD = TDMaize, trait = "yld")
## Calculate new mega-environments based on the genotypes with the lowest
## value per environment.
TDmegaEnv2 <- gxeMegaEnv(TD = TDmegaEnv, trait = "yld", method = "min")

}
