% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeStability.R
\name{gxeStability}
\alias{gxeStability}
\title{Calculates stability coefficients for genotype-by-environment data}
\usage{
gxeStability(TD, trait, method = c("superiority", "static", "wricke"),
  bestMethod = c("max", "min"), sorted = c("descending", "ascending",
  "none"))
}
\arguments{
\item{TD}{an object of class \code{\link{TD}}.}

\item{trait}{A character string specifying the trait to be analyzed.}

\item{method}{A character vector specifying (a) measure(s) of stability to be
calculated. Options are superiority (cultivar-superiority measure), static
(Shukla's stability variance) or wricke (wricke's ecovalence). By default
all three measures are calculated.}

\item{bestMethod}{A character string specifying the criterion to define
the best genotype ("max","min").}

\item{sorted}{A character string specifying the sorting order of the results.}
}
\value{
A list of one to three data.frames containing the stability measures.
}
\description{
This function calculate difference measures of stability, such as the
cultivar-superiority measure of Lin & Binns (1988), Shukla's (1972) stability
variance and Wricke's (1962) ecovalence.
}
\examples{
geStab <- gxeStability(TD = TDMaize, trait = "yld", sorted = "descending")
report(geStab, outfile = "./testReports/reportStability.pdf")

}
\references{
LiN, C. S. and Binns, M. R. 1988. A superiority measure of cultivar
performance for cultivar x location data. Can. J. Plant Sci. 68: 193-198

Shukla, G.K. 1972. Some statistical aspects of partitioning
genotype-environmental components of variability. Heredity 29:237-245

Wricke, G. Uber eine method zur erfassung der okologischen
streubreit in feldversuchen. Zeitschrift für Pflanzenzucht,
v. 47, p. 92-96, 1962
}
