% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeTable.R
\name{gxeTable}
\alias{gxeTable}
\title{Compute BLUPS and based on a set of mega-environments}
\usage{
gxeTable(TD, trait, useYear = FALSE, engine = c("asreml", "lme4"), ...)
}
\arguments{
\item{TD}{an object of class \code{\link{TD}}.}

\item{trait}{A character string specifying the trait to be analyzed.}

\item{useYear}{Should year be used for modelling (as years within
environments). If \code{TRUE} TD should contain a column "year".}

\item{engine}{A character string specifying the name of the mixed engine to
use, either lme4 or asreml.}

\item{...}{Other parameters passed to either \code{asreml} or \code{lmer}.}
}
\value{
A list consisting of two data.frames, \code{predictedValue}
containing BLUPs per genotype per mega-environment and \code{standardError}
containing standard errors for those BLUPs.
}
\description{
This function calculates predicted means (BLUPS) and associated standard
errors based on a set of mega-environments.
}
\examples{
TDMegaEnv <- gxeMegaEnv(TD = TDMaize, trait = "yld", sumTab = FALSE)
geTab <- gxeTable(TD = TDMegaEnv, trait = "yld")

}
