% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeVarComp.R
\name{gxeVarComp}
\alias{gxeVarComp}
\title{Selects the best variance-covariance model for a set of trials}
\usage{
gxeVarComp(
  TD,
  trials = names(TD),
  trait,
  engine = c("lme4", "asreml"),
  trialGroup = NULL,
  useWt = FALSE,
  ...
)
}
\arguments{
\item{TD}{An object of class \code{\link{TD}}.}

\item{trials}{A character string specifying the trials to be analyzed. If
not supplied, all trials are used in the analysis.}

\item{trait}{A character string specifying the trait to be analyzed.}

\item{engine}{A character string specifying the engine used for modeling.
Either "lme4" or "asreml".}

\item{trialGroup}{A character string specifying a column in TD.......}

\item{useWt}{Should weighting be used when modeling? Requires a column
\code{wt} in \code{TD}.}

\item{...}{Further arguments to be passed to \code{asreml}.}
}
\value{
An object of class \code{\link{varComp}}, a list object containing:
\item{STA}{An object of class STA containing the best fitted model.}
\item{choice}{A character string indicating the best fitted model.}
\item{summary}{A data.frame with a summary of the fitted models.}
\item{engine}{A character string containing the engine used for
the analysis.}
}
\description{
This function selects the best covariance structure for genetic correlations
between trials. It fits a range of variance-covariance models (identity,
compound symmetry (cs), diagonal, simple correlation with heterogeneous
variance (outside), heterogeneous compound symmetry (hcs),
first order factor analytic (fa), second order factor analytic (fa2) and
unstructured), and selects the best one using a goodness-of-fit criterion.
}
\note{
If \code{engine = "lme4"}, only the compound symmetry model can be
fitted.
}
\examples{
## Select the best variance-covariance model using lme4 for modeling.
geVarComp <- gxeVarComp(TD = TDMaize, trait = "yld")

## Summarize results.
summary(geVarComp)

\dontrun{
## Create a pdf report summarizing the results.
report(geVarComp, outfile = "./testReports/reportVarComp.pdf")
}

\dontrun{
## Select the best variance-covariance model using asreml for modeling.
## Use BIC as a goodness-of-fit criterion.
geVarComp2 <- gxeVarComp(TD = TDMaize, trait = "yld", engine = "asreml",
                        criterion = "BIC")

summary(geVarComp2)

## Plot a heatmap of the correlation matrix for the best model.
plot(geVarComp2)
}
}
