% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kurtosis}
\alias{kurtosis}
\title{Helper function for computing kurtosis.
Rescaled by subtracting 3 from the result to give the normal distribution
a kurtosis of 0, so basically the excess kurtosis.}
\usage{
kurtosis(x, na.rm = FALSE)
}
\description{
Helper function for computing kurtosis.
Rescaled by subtracting 3 from the result to give the normal distribution
a kurtosis of 0, so basically the excess kurtosis.
}
\keyword{internal}
