% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMegaEnv.R
\name{megaEnv}
\alias{megaEnv}
\alias{createMegaEnv}
\title{S3 class megaEnv}
\usage{
createMegaEnv(TD, summTab, trait)
}
\arguments{
\item{TD}{A data.frame containing values for the
cultivar-superiority measure of Lin and Binns.}

\item{summTab}{A data.frame containing values for Shukla's stability variance.}

\item{trait}{A character string indicating the trait that has been analyzed.}
}
\description{
Function for creating objects of S3 class megaEnv.\cr
\code{\link{print}}, \code{\link{summary}}, \code{\link{plot}} and
\code{\link{report}} methods are available.
}
\seealso{
\code{\link{predict.megaEnv}}
}
\keyword{internal}
