% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultiQTL.R
\name{multiQTL}
\alias{multiQTL}
\alias{createMultiQTL}
\title{S3 class multiQTL}
\usage{
createMultiQTL(qtl, QTLDet, selection, thr)
}
\arguments{
\item{qtl}{A fitted multi QTL model.}

\item{QTLDet}{The object of class \code{\link{QTLDet}} used as base for
fitting the QTL model.}

\item{selection}{A character string indictating the type of selection used
for selecting the markers in the final model.}

\item{thr}{A numerical value indicating the threshold for dropping terms in
the backwards elemination process.}
}
\description{
Function for creating objects of S3 class multiQTL.\cr
\code{\link{summary}}, \code{\link{plot}} and \code{\link{report}}
methods are available.
}
\seealso{
\code{\link{plot.multiQTL}}, \code{\link{report.multiQTL}}
}
\author{
Bart-Jan van Rossum
}
\keyword{internal}
