% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlierTD}
\alias{outlierTD}
\title{Identifying Outliers}
\usage{
outlierTD(TD, traits, coef = 1.5, commonFactors = NULL)
}
\arguments{
\item{TD}{An object of class \code{\link{TD}}.}

\item{traits}{A character vector specifying the name(s) of the traits for
which outliers should be identified.}

\item{coef}{A numerical value used for determining when a value is
considered an outlier. All observations that exceed \code{coef} times
the interquartile range will be marked as outliers.}

\item{commonFactors}{A character vector specifying the names of columns
in \code{TD} used for selecting observations that are similar to the
outliers. If \code{commonFactors = NULL} only outliers are reported and
no similar observations.}
}
\value{
A data.frame containing logical values indicating if the
observation is an outlier.
}
\description{
Function to identify observations that exceed \code{coef} times the
interquartile range. A summary is printed of outliers and observations
that have the same value for \code{commonFactors}. The latter ones will
be marked as similar to distinguish them from the former ones.
}
\examples{
outliers <- outlierTD(TD = TDHeat05, traits = "yield")

}
