% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAMMI.R
\name{plot.AMMI}
\alias{plot.AMMI}
\title{Plot function for class AMMI}
\usage{
\method{plot}{AMMI}(x, ..., plotType = c("AMMI1", "AMMI2"), scale = 1,
  col = c("orange3", "navyblue"))
}
\arguments{
\item{x}{An object of class AMMI}

\item{...}{Further graphical parameters passed on to actual plot function.}

\item{plotType}{A character string indicating which plot should be made.
Either "AMMI1" or "AMMI2" for an AMMI1 biplot (genotype and
environment means vs PC1) or an AMMI2 biplot (genotypes and environment
interaction with PC1 and PC2) respectively.}

\item{scale}{A numerical value. The variables are scaled by
\code{lambda ^ scale} and the observations by \code{lambda ^ (1 - scale)}
where \code{lambda} are the singular values computed by
\code{\link[stats]{princomp}} in \code{\link{gxeAmmi}}. Normally
\code{0 <= scale <= 1}, and a warning will be issued if the specified
scale is outside this range.}

\item{col}{A character vector with plot colors for genotype and environment.}
}
\value{
A biplot depending on \code{plotType}.
}
\description{
Two types of biplot can be made. A biplot of genotype and environment
means vs PC1 (AMMI1) or a biplot of genotypes and environment interaction
with PC1 and PC2 (AMMI2).
}
\examples{
## Run AMMI analysis.
geAmmi <- gxeAmmi(TD = TDMaize, trait = "yld")
## Create a biplot of genotype and environment means vs PC1.
plot(geAmmi)
## Create a biplot of genotypes and environment interaction with PC1 and PC2.
plot(geAmmi, plotType = "AMMI2")

}
