% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFW.R
\name{plot.FW}
\alias{plot.FW}
\title{Plot Function for Class FW}
\usage{
\method{plot}{FW}(x, ..., plotType = c("scatter", "line", "trellis"),
  sorted = c("ascending", "descending", "none"))
}
\arguments{
\item{x}{An object of class FW}

\item{...}{Other graphical parameters passed on to actual plot function.}

\item{plotType}{A character string indicating which plot should be made.
Possible values are "scatter", "line"  and "trellis" for creating a scatter
plot of sensitivities, a plot of fitted lines for each genotype and a trellis
plot of the individual genotype slopes respectively.}

\item{sorted}{A character string specifying whether the results are to be sorted
in an increasing (or decreasing) order of sensitivities.}
}
\value{
A plot depending on \code{plotType}
}
\description{
Three types of plot can be made. A scatter plot for genotypic mean,
mse and sensitivity, a line plot with fitted lines for each genotype and
a trellis plot with individual slopes per genotype (for max 64 genotypes).
}
\examples{
# Run Finlay-Wilkinson analysis.
geFW <- gxeFw(TD = TDMaize, trait = "yld")
# Create scatter plot.
plot(geFW, plotType = "scatter")

}
