% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMegaEnv.R
\name{plot.megaEnv}
\alias{plot.megaEnv}
\title{Plot function for class megaEnv}
\usage{
\method{plot}{megaEnv}(x, ..., output = TRUE)
}
\arguments{
\item{x}{An object of class megaEnv.}

\item{...}{Further arguments to be passed on to underlying plot functions.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a ggplot object is invisibly returned.}
}
\description{
Function for creating scatter plots of predicted values in computed mega
environments.
}
