% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultiQTL.R
\name{plot.multiQTL}
\alias{plot.multiQTL}
\title{Plot function for class multiQTL}
\usage{
\method{plot}{multiQTL}(x, ...,
  main = "QTL estimates and confidence intervals")
}
\description{
Plotting function for objects of class multiQTL. Plots the estimates of the
QTLs in the final model with their respective confidence intervals.
}
\details{
@param x An object of class multiQTL
 @param ... Further graphical parameters. Currently not used.
 @param main A character string used as overall title for the plot.
}
