% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStability.R
\name{plot.stability}
\alias{plot.stability}
\title{Plot Function for Class stability}
\usage{
\method{plot}{stability}(x, ...)
}
\arguments{
\item{x}{an object of class stability.}

\item{...}{other arguments to be passed on to underlying plot functions.}
}
\value{
Plots of stability measures against means.
}
\description{
Function for creating plots of computed stability measures against the means.
}
