% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarComp.R
\name{plot.varComp}
\alias{plot.varComp}
\title{Plot function for class varComp}
\usage{
\method{plot}{varComp}(x, ...)
}
\arguments{
\item{x}{An object of class varComp}

\item{...}{Not used}
}
\description{
Function for plotting a heatmap of the correlation matrix for objects of
class varComp.
}
\examples{
\dontrun{
## Select the best variance-covariance model using asreml for modeling.
geVarComp <- gxeVarComp(TD = TDMaize, trait = "yld", engine = "asreml")
## Create a heatmap of the correlation matrix for the best model.
plot(geVarComp)
}

}
