% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAMMI.R
\name{plotAMMI1}
\alias{plotAMMI1}
\title{Helper function for creating AMMI1 plot}
\usage{
plotAMMI1(loadings, scores, importance, overallMean, genoMean, envMean,
  trait, dat, GGE, year = "", scale, colGeno, colEnv, colorBy)
}
\description{
Helper function for creating AMMI1 plot
}
\keyword{internal}
