% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAMMI.R
\name{plotAMMI2}
\alias{plotAMMI2}
\title{Helper function for creating AMMI2 plot}
\usage{
plotAMMI2(loadings, scores, importance, trait, year = "",
  primAxis = "PC1", secAxis = "PC2", scale, col)
}
\description{
Helper function for creating AMMI2 plot
}
\keyword{internal}
