% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotCorMat}
\alias{plotCorMat}
\title{Function for plotting a correlation (or covariance) matrix.}
\usage{
plotCorMat(corMat, main = "")
}
\description{
Function for plotting a correlation (or covariance) matrix.
}
\keyword{internal}
