% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predictAsreml}
\alias{predictAsreml}
\title{Extended version of asreml.predict}
\usage{
predictAsreml(model, classify = "genotype",
  associate = as.formula("~ NULL"), vcov = TRUE, TD, ...)
}
\description{
Asreml has a bug that may throw a warning message:
Abnormal termination
Insufficient workspace - (reset workspace or pworkspace arguments)
This may be avoided by increasing pworkspace, but this doesn't
always work.
If this happens pworkspace is increased in 'small' steps.
}
\keyword{internal}
