% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSSA.R
\name{print.summary.SSA}
\alias{print.summary.SSA}
\title{Printing summazed objects of class SSA}
\usage{
\method{print}{summary.SSA}(x, digits = max(getOption("digits") - 2, 3), ...)
}
\arguments{
\item{x}{An object of class \code{summary.SSA}}

\item{digits}{An integer indicating the number of significant digits for
printing.}

\item{...}{Further arguments passed to \code{\link[stats]{printCoefmat}}.}
}
\description{
\code{print} method for object of class summary.SSA created by summarizing
objects of class SSA.
}
