% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printAnova}
\alias{printAnova}
\title{Helper function for printing anova table in reports.}
\usage{
printAnova(aovTab, title = NULL)
}
\description{
Helper function for printing anova table in reports.
}
\keyword{internal}
