% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLDet.R
\name{QTLDet}
\alias{QTLDet}
\alias{createQTLDet}
\title{S3 class QTLDet}
\usage{
createQTLDet(scores, peaks, type, cross, trait, info)
}
\arguments{
\item{scores}{A data.frame containing the lod scores.}

\item{peaks}{A data.frame containing the peaks found.}

\item{type}{A character string indicating the type of QTLDetection performed.}

\item{cross}{An object of class cross in the \code{qtl} package.}

\item{trait}{A character string indicating the trait for which the analysis
is done.}

\item{info}{A list containing information on the settings used for
QTL Detection, step, threshold and window.}
}
\description{
Function for creating objects of S3 class QTLDet.\cr
\code{\link{print}}, \code{\link{summary}}, \code{\link{plot}} and
\code{\link{report}} methods are available.
}
\seealso{
\code{\link{plot.QTLDet}}, \code{\link{report.QTLDet}}
}
\author{
Bart-Jan van Rossum
}
