% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTLMapQC.R
\name{QTLMapQC}
\alias{QTLMapQC}
\title{Quality Control and Cleaning of a cross object}
\usage{
QTLMapQC(cross, missMrk = 0.05, missInd = 0.05,
  removeDuplicates = TRUE, segDistortion = 0.001, recombination = 3,
  reestimateMap = FALSE, crossover = 0.2)
}
\arguments{
\item{cross}{An object of class cross created by the qtl package.}

\item{missMrk}{A numerical value between 0 and 1 indicating the maximum
allowed fraction of missing values per marker. Markers with a fraction of
missing values above \code{missMrk} will be removed.}

\item{missInd}{A numerical value between 0 and 1 indicating the maximum
allowed fraction of missing values per individual. Individuals with a fraction of
missing values above \code{missMrk} will be removed.}

\item{removeDuplicates}{Should duplicate markers be removed?}

\item{segDistortion}{A numerical value between 0 and 1 used a threshold for
Mendelian segregation. Markers with a P-value below \code{segDistortion} will
be removed.}

\item{recombination}{A positive numerical value used a threshold for checking
recombination between pairs of markers.}

\item{reestimateMap}{Should the map be reestimated based on the observed markers?}

\item{crossover}{A numerical value between 0 and 1 indicating the maximum
allowed fraction of crossovers per individual. Individuals with a fraction of
crossovers above \code{crossover} will be removed.}
}
\value{
A cleaned version of the input cross object after markers and
individuals have been removed that are outside the respective thresholds.
}
\description{
Function for performing quality control and cleaning of a cross object.
Quality control is done in several subsequent steps:
\enumerate{
\item{Markers with a fraction of missing values higher than \code{missMrk}
are removed.}
\item{Duplicated markers are removed.}
\item{Individuals with a fraction of missing values higher than \code{missInd}
are removed.}
\item{Markers that show evidence of segregation distortion (P-value below
\code{segDistortion}) are removed.}
\item{Markers which might have been switched (with threshold
\code{recombination}) are removed. See also
\code{\link[qtl]{checkAlleles}}.}
\item{The map is reestimated based on the observed markers.}
\item{Individuals with a fraction of crossovers higher than
\code{crossover} are removed.}
}
Steps 1, 3, 4, 5 and 7 are only performed if their respective threshold
values are positive. Setting them to 0 suppresses the corresponding check.\cr
Steps 2 and 6 are performed if respectively \code{removeDuplicates} and
\code{reestimateMap} are \code{TRUE}.
}
\examples{
## Read the data.
F2 <- qtl::read.cross(format="csv",
                      file = system.file("extdata", "F2maize_geno.csv",
                                         package = "RAP"),
                      genotypes = c("AA", "AB", "BB"),
                      alleles = c("A", "B"), estimate.map = FALSE)
## Run quality control.
F2QC <- QTLMapQC(F2)
## Compare cross object before and after cleaning.
summary(F2)
summary(F2QC)

## Run quality control: only remove markers with a fraction of missing
## values higher than 0.02
F2QC2 <- QTLMapQC(F2, missMrk = 0.02, missInd = 0, removeDuplicates = FALSE,
                 segDistortion = 0, recombination = 0, crossover = 0)
summary(F2QC2)

}
\references{
Broman et al. (2003) R/qtl: QTL mapping in experimental crosses.
Bioinformatics 19:889-890
}
\seealso{
\code{\link[qtl]{nmissing}}, \code{\link[qtl]{drop.markers}},
\code{\link[qtl]{drop.dupmarkers}}, \code{\link[qtl]{geno.table}},
\code{\link[qtl]{nmissing}}, \code{\link[qtl]{checkAlleles}},
\code{\link[qtl]{replace.map}}, \code{\link[qtl]{countXO}}
}
