% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qtlPosToName}
\alias{qtlPosToName}
\title{Function for extracting the name of a qtl combining the peaks from a QTLDet
object and the position of the qtl on the chromosome.}
\usage{
qtlPosToName(chrPos, peaks)
}
\description{
Function for extracting the name of a qtl combining the peaks from a QTLDet
object and the position of the qtl on the chromosome.
}
\keyword{internal}
