% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{createTD}
\alias{fitTD}
\alias{STAtoTD}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{statgenSTA}{\code{\link[statgenSTA]{createTD}}, \code{\link[statgenSTA]{fitTD}}, \code{\link[statgenSTA]{STAtoTD}}}
}}

