% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAMMI.R
\name{report.AMMI}
\alias{report.AMMI}
\title{Report method for class AMMI}
\usage{
\method{report}{AMMI}(x, ..., outfile = NULL)
}
\arguments{
\item{x}{An object of class AMMI.}

\item{...}{Further arguments passed on from other functions - not used yet.}

\item{outfile}{A character string, the name and location of the output .pdf and .tex
file for the report. If \code{NULL} a report will be created in the current working
directory.}
}
\description{
A pdf report will be created containing a summary of AMMI model.
Simultaneously the same report will be created as a tex file.
}
