% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTLMapQC.R
\name{report.cross}
\alias{report.cross}
\title{Report method for class cross}
\usage{
\method{report}{cross}(x, ..., outfile = NULL)
}
\arguments{
\item{x}{An object of class cross.}

\item{...}{Further arguments passed on from other functions - not used yet.}

\item{outfile}{A character string, the name and location of the output .pdf
and .tex file for the report. If \code{NULL} a report with a default name
will be created in the current working directory.}
}
\description{
A pdf report will be created containing a summary of a cross object.
Simultaneously the same report will be created as a tex file.
}
\examples{
## Read the data
F2 <- qtl::read.cross(format="csv",
                      file = system.file("extdata",
                                        "F2_maize_practical3_ex2.csv",
                                        package = "RAP"),
                      genotypes = c("AA", "AB", "BB"),
                      alleles = c("A", "B"), estimate.map = FALSE)
\dontrun{
## Create a report
report(F2, outfile = "./testReports/reportCross.pdf")
}

}
