% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultiQTL.R
\name{report.multiQTL}
\alias{report.multiQTL}
\title{Report method for class multiQTL}
\usage{
\method{report}{multiQTL}(x, ..., outfile = NULL)
}
\arguments{
\item{x}{An object of class multiQTL.}

\item{...}{Further arguments passed on from other functions - not used yet.}

\item{outfile}{A character string, the name and location of the output .pdf
and .tex file for the report. If \code{NULL}, a report with a default name
will be created in the current working directory.}
}
\value{
A pdf and tex report.
}
\description{
A pdf report will be created containing a summary of a multiQTL analysis.
Simultaneously the same report will be created as a tex file.
}
\examples{
## Read the data
F2 <- qtl::read.cross(format="csv",
                      file = system.file("extdata", "F2maize_geno.csv",
                                        package = "RAP"),
                      genotypes = c("AA", "AB", "BB"),
                      alleles = c("A", "B"), estimate.map = FALSE)
## Perform QTL detection using simple interval mapping.
QTLDet <- QTLDetect(cross = F2, trait = "trait", type = "SIM")
## Fit a multi QTL model.
multiFit <- multiQTLFit(QTLDet)
\dontrun{
## Create a pdf report summarizing results.
report(multiFit, outfile = "./testReports/reportMultiQTLFit.pdf")
}

}
