% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLDet.R
\name{report.QTLDet}
\alias{report.QTLDet}
\title{Report method for class QTLDet}
\usage{
\method{report}{QTLDet}(x, ..., outfile = NULL)
}
\arguments{
\item{x}{an object of class QTLDet.}

\item{...}{further arguments passed on from other functions - not used yet.}

\item{outfile}{a character string, the name and location of the output .pdf and .tex
file for the report. If \code{NULL} a report will be created in the current working
directory.}
}
\description{
A pdf report will be created containing a summary of QTLDet analysis.
Simultaneously the same report will be created as a tex file.
}
