% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report}
\alias{report}
\title{Base method for creating a report}
\usage{
report(x, ...)
}
\arguments{
\item{x}{An \code{R} object}

\item{...}{Further arguments to be passed on to specific report functions.}
}
\description{
Base method for creating a .pdf and .tex report from an \code{R} object.
}
\seealso{
\code{\link{report.varComp}}, \code{\link{report.AMMI}},
\code{\link{report.FW}}, \code{\link{report.stability}}
}
