% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSSA.R
\name{report.SSA}
\alias{report.SSA}
\title{Report method for class SSA}
\usage{
\method{report}{SSA}(x, ..., descending = TRUE, outfile = NULL, what = if
  (is.null(x$mFix)) "random" else "fixed")
}
\arguments{
\item{x}{An object of class SSA.}

\item{...}{Further arguments passed on from other functions - not used yet.}

\item{descending}{Should the trait be ordered in descending order? Set to
\code{FALSE} if low values of the trait indicate better performance.}

\item{outfile}{A character string, the name and location of the output .pdf and .tex
file for the report. If \code{NULL} a report will be created in the current working
directory.}

\item{what}{A character string indicating whether the model with genotype fixed
or genotype random should be reported.}
}
\description{
A pdf report will be created containing a summary of the results of the model.
Simultaneously the same report will be created as a tex file.
}
