% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStability.R
\name{report.stability}
\alias{report.stability}
\title{Report method for class stability}
\usage{
\method{report}{stability}(x, ..., outfile = NULL)
}
\arguments{
\item{x}{an object of class stability.}

\item{...}{further arguments passed on from other functions - not used yet.}

\item{outfile}{a character string, the name and location of the output .pdf and .tex
file for the report. If \code{NULL} a report will be created in the current working
directory.}
}
\description{
A pdf report will be created containing a summary of stability model.
Simultaneously the same report will be created as a tex file.
}
