% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarComp.R
\name{report.varComp}
\alias{report.varComp}
\title{Report method for class varComp}
\usage{
\method{report}{varComp}(x, ..., outfile = NULL)
}
\arguments{
\item{x}{An object of class varComp.}

\item{...}{Further arguments passed on from other functions - not used yet.}

\item{outfile}{A character string, the name and location of the output .pdf
and .tex file for the report. If \code{NULL}, a report with a default name
will be created in the current working directory.}
}
\value{
A pdf and tex report.
}
\description{
A pdf report will be created containing a summary of an object of class
varComp. Simultaneously the same report will be created as a tex
file.
}
