% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAMMI.R
\name{residuals.AMMI}
\alias{residuals.AMMI}
\title{Extract residuals.}
\usage{
\method{residuals}{AMMI}(object, ...)
}
\arguments{
\item{object}{An object of class AMMI}

\item{...}{Not used.}
}
\description{
Extract the residuals for the fitted AMMI model.
}
