% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFW.R
\name{residuals.FW}
\alias{residuals.FW}
\title{Extract residuals.}
\usage{
\method{residuals}{FW}(object, ...)
}
\arguments{
\item{object}{An object of class FW}

\item{...}{Not used.}
}
\description{
Extract the residuals for the fitted FW model.
}
