% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarCov.R
\name{residuals.varCov}
\alias{residuals.varCov}
\title{Extract residuals.}
\usage{
\method{residuals}{varCov}(object, ...)
}
\arguments{
\item{object}{An object of class varCov}

\item{...}{Not used.}
}
\description{
Extract the residuals for the best model.
}
