% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STExtract.R
\name{restoreColNames}
\alias{restoreColNames}
\title{Helper function for adding back original colnames}
\usage{
restoreColNames(renDat, renamedCols, restore = FALSE)
}
\description{
Helper function for adding back original colnames
}
\keyword{internal}
