% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seKurtosis}
\alias{seKurtosis}
\title{Helper function for computing the standard error of the kurtosis.}
\usage{
seKurtosis(n)
}
\description{
Helper function for computing the standard error of the kurtosis.
}
\keyword{internal}
