% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seSkewness}
\alias{seSkewness}
\title{Helper function for computing the standard error of the skewness.}
\usage{
seSkewness(n)
}
\description{
Helper function for computing the standard error of the skewness.
}
\keyword{internal}
