% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTD.R
\name{setMeta}
\alias{setMeta}
\title{Set metadata of TD objects}
\usage{
setMeta(TD, meta)
}
\arguments{
\item{TD}{An object of class TD.}

\item{meta}{A data.frame containing metadata.}
}
\value{
The object of class TD with updated metadata.
}
\description{
Function for setting metadata of a TD object for one or more trials
simultaneously. Metadata can be set using a data.frame with rownames
corresponding to the trials in \code{TD}. The data.frame should contain one
or  more of the following columns: trLocation, trDate, trDesign, trLat,
trLong, trPlWidth and trPlLength. The values of the metadata of TD
will be set to the values in the corresponding column in \code{meta}.
Existing values will be overwritten, but \code{NA} will be ignored so
setting a value to \code{NA} won't result in accidentally removing it.
}
\seealso{
Other functions for TD objects: \code{\link{TD}},
  \code{\link{getMeta}}, \code{\link{plot.TD}},
  \code{\link{summary.TD}}
}
\concept{functions for TD objects}
