% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seVar}
\alias{seVar}
\title{Helper function for computing the standard error of the variance.}
\usage{
seVar(x, na.rm = FALSE)
}
\description{
Helper function for computing the standard error of the variance.
}
\keyword{internal}
