% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skewness}
\alias{skewness}
\title{Helper function for computing the skewness.
This and following formulas taken from
https://brownmath.com/stat/shape.htm#Normal.}
\usage{
skewness(x, na.rm = FALSE)
}
\description{
Helper function for computing the skewness.
This and following formulas taken from
https://brownmath.com/stat/shape.htm#Normal.
}
\keyword{internal}
