% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSSA.R
\name{SSA}
\alias{SSA}
\alias{createSSA}
\title{S3 class SSA}
\usage{
createSSA(mRand, mFix, TD, traits = NULL, design = NULL, spatial = NULL,
  engine = NULL, predicted = "genotype")
}
\arguments{
\item{mRand}{a model with genotype random.}

\item{mFix}{a model with genotype fixed.}

\item{TD}{an object of class \code{\link{TD}} containing the data on which
\code{mRand} and \code{mFix} are based.}

\item{traits}{a character vector indicating the traits for which the analysis
is done.}

\item{design}{a character string containing the design of the trial.
(see \code{\link{STRunModel}} for the possible designs).}

\item{spatial}{a character string indicating the spatial part of the model.
\code{FALSE} if no spatial design has been used.}

\item{engine}{a character string containing the engine used for the analysis.}

\item{predicted}{a character string indicating the variable that has been predicted.}
}
\description{
Function for creating objects of S3 class Single Site Analysis (SSA).\cr
\code{\link{summary}}, \code{\link{plot}} and \code{\link{report}}
methods are available.
}
\seealso{
\code{\link{summary.SSA}}, \code{\link{plot.SSA}}, \code{\link{report.SSA}}
}
\author{
Bart-Jan van Rossum
}
