% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSSA.R
\name{SSAtoTD}
\alias{SSAtoTD}
\title{Convert SSA to TD}
\usage{
SSAtoTD(SSA, what = c("BLUEs", "seBLUEs", "BLUPs", "seBLUPs"),
  traits = NULL, keep = NULL, addWt = FALSE)
}
\arguments{
\item{SSA}{An object of class \code{\link{SSA}}.}

\item{what}{A character string containing the statistics to be included as
traits in the TD object. Multiple statistics can be included in which case
they will appear as \code{statistic_trait} in the output}

\item{traits}{A character string containing the traits to be included in the
TD object. If \code{NULL}, all traits are exported.}

\item{keep}{Columns from the TD object used as input for the SSA model to
be copied to the output. see \code{\link{extract}} for possible columns to
copy. If if it is available in \code{TD}, the column \code{trial} will always
be copied.}

\item{addWt}{Should a column wt be added to the output? If \code{TRUE}
weight is calculated as 1/(SE BLUEs) ^ 2. If multiple traits are included in
the output, multiple weight columns will be added, 1 for each trait. These
will be named \code{wt_trait}.}
}
\description{
Convert an SSA object to a TD object.\cr
To be able to use the output of a single site analysis in
Genotype-by-Environment (GxE) analysis the output first needs to be converted
back to an TD object. This function does exactly that. It extracts BLUEs,
BLUPs and their standard errors from the SSA object and creates a new TD
object using these. Also a column "wt" (weigth) may also be added. Weights
are then calculated as 1/(SE BLUEs) ^ 2.
}
\details{
Trial information for the trials in the SSA object will be copied from the
original TD object on which the modeling was done.
}
\examples{
## Run model using SpATS.
myModel <- fitTD(TD = TDHeat05, design = "res.rowcol", traits = "yield",
                      what = "fixed")
## Create TD object from the fitted model.
myTD <- SSAtoTD(myModel)

}
