% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenGxE.R
\docType{package}
\name{statgenGxE-package}
\alias{statgenGxE}
\alias{statgenGxE-package}
\title{statgenGxE: Genotype by Environment (GxE) Analysis}
\description{
Analysis of multi environment data of plant breeding experiments following the analyses described in Malosetti, Ribaut, and van Eeuwijk (2013), \doi{10.3389/fphys.2013.00044}. One of a series of statistical genetic packages for streamlining the analysis of typical plant breeding experiments developed by Biometris. Some functions have been created to be used in conjunction with the R package 'asreml' for the 'ASReml' software, which can be obtained upon purchase from 'VSN' international (\url{https://vsni.co.uk/software/asreml-r/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biometris.github.io/statgenGxE/index.html}
  \item \url{https://github.com/Biometris/statgenGxE/}
  \item Report bugs at \url{https://github.com/Biometris/statgenGxE/issues}
}

}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})

Other contributors:
\itemize{
  \item Fred van Eeuwijk (\href{https://orcid.org/0000-0003-3672-2921}{ORCID}) [contributor]
  \item Martin Boer (\href{https://orcid.org/0000-0002-1879-4588}{ORCID}) [contributor]
  \item Marcos Malosetti (\href{https://orcid.org/0000-0002-8150-1397}{ORCID}) [contributor]
  \item Daniela Bustos-Korts (\href{https://orcid.org/0000-0003-3827-6726}{ORCID}) [contributor]
  \item Emilie J Millet (\href{https://orcid.org/0000-0002-2913-4892}{ORCID}) [contributor]
  \item Joao Paulo (\href{https://orcid.org/0000-0002-4180-0763}{ORCID}) [contributor]
  \item Maikel Verouden (\href{https://orcid.org/0000-0002-4893-3323}{ORCID}) [contributor]
  \item Willem Kruijer (\href{https://orcid.org/0000-0001-7179-1733}{ORCID}) [contributor]
  \item Ron Wehrens (\href{https://orcid.org/0000-0002-8798-5599}{ORCID}) [contributor]
  \item Choazhi Zheng (\href{https://orcid.org/0000-0001-6030-3933}{ORCID}) [contributor]
  \item Turbet Delof Michel [contributor]
  \item Rameau Jean-Francois [contributor]
}

}
\keyword{internal}
