% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STExtract.R
\name{STExtract}
\alias{STExtract}
\title{Extract statistics from Fitted Models}
\usage{
STExtract(SSA, traits = SSA$traits, what = "all", keep = NULL)
}
\arguments{
\item{SSA}{An object of class SSA.}

\item{traits}{A character vector of traits for which the statistics should be
computed If not supplied statistics are computed for all traits that have
been modelled.}

\item{what}{A character vector indicating which statistics should be computed.
Most statistics are available for all models, some only for models fitted using
a certain engine. If this is the case this is indicated in the list with options
in details.\cr
If \code{what = "all"} all available statistics are computed.\cr}

\item{keep}{A character vector of column(s) in the object of class
\code{\link{TD}} used for modeling. These columns will be kept as output when
computing fitted values, residuals, standardized residuals and rMeans. Columns
can also be kept when computing (se)BLUEs and (se)BLUPs but only if the column
to keep contains unique values for the modeled variables, i.e. a column repId
with several values per genotype cannot be kept.}
}
\value{
A list of extracted statistics.
}
\description{
This function extracts and calculates various results for fitted models such as
BLUEs, BLUPs, unit errors, heritabilities. Note that most results can only
calculated if a model is fitted with genotype is fixed or random. This is
indicated in the list below with "F" and "R"
}
\details{
Possible options for \code{what} are:
\describe{
\item{F - BLUEs}{Best Lineair Unbiased Estimatiors}
\item{F - seBLUES}{Standard errors of the BLUEs}
\item{R - BLUPs}{Best Lineair Unbiased Predictors}
\item{R - seBLUPs}{Standard errors of the BLUPs}
\item{F - ue}{unit errors - only for \code{lme4} and \code{asreml}}
\item{R - heritability}{heritability}
\item{R - varGen}{genetic variance component}
\item{R - varErr}{residual variance component - only for \code{lme4}
and \code{asreml}}
\item{R - varSpat}{spatial variance components - only for \code{SpATS}}
\item{F - fitted}{fitted values for the model with genotype as fixed component}
\item{F - resid}{residuals for the model with genotype as fixed component}
\item{F - stdRes}{standardized residuals for the model with genotype as fixed
component - only for \code{lme4} and \code{asreml}}
\item{R - rMeans}{fitted values for the model with genotype as random component}
\item{R - ranEf}{random genetic effects}
\item{F - wald}{results of the wald test}
\item{F - CV}{coefficient of variation - only for \code{lme4} and \code{asreml}}
\item{F - rDf}{residual degrees of freedom}
\item{R - effDim}{effective dimensions - only for \code{SpATS}}
\item{F - sed}{standard error of difference - only for \code{asreml}}
\item{F - lsd}{least significant difference - only for \code{asreml}}
}
}
\examples{
## Fit model using lme4.
myModel <- STRunModel(TD = TDHeat05, design = "res.rowcol", traits = "yield")

## Extract statistics from fitted model.
extr <- STExtract(myModel)

}
\seealso{
\code{\link{STRunModel}}, \code{\link{STModSpATS}}, \code{\link{STModLme4}} and
\code{\link{STModAsreml}}
}
