% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STModLme4.R
\name{STModLme4}
\alias{STModLme4}
\title{Fit Single Trial Model using lme4}
\usage{
STModLme4(TD, trials = names(TD), traits, what = c("fixed", "random"),
  covariates = NULL, useCheckId = FALSE, control = NULL,
  trySpatial = FALSE, design = "rowcol", checks = TRUE, ...)
}
\arguments{
\item{TD}{An object of class \code{\link{TD}}.}

\item{traits}{A character vector specifying the traits for modeling.}

\item{what}{A character vector specifying whether "genotype" should
be fitted as "fixed" or "random" effect. If not specified both models
are fitted.}

\item{covariates}{A character vector specifying covariates to be fitted as
extra fixed effects in the model.}

\item{useCheckId}{Should checkId be used as a fixed effect in the model?\cr
If \code{TRUE} \code{TD} has to contain a column 'checkId'.}

\item{control}{An optional list with control parameters to be passed to the
actual fitting funcions. Currently \code{nSeg} and \code{nestDiv} are valid
parameters when fitting a model using SpATS. They pass a value to nseg and
nest.div in \code{\link[SpATS]{PSANOVA}} respectively. \code{criterion} is a
valid parameter when fitting a spatial model using asreml. Use this to pass
a goodness-of-fit criterion for comparing different spatial models. See also
in details. Other parameters are ignored.}

\item{trySpatial}{Should spatial models be tried? Spatial models can
only be fitted with SpATS and asreml. If SpATS is used for modeling only
spatial models can be fitted and trySpatial is always set to \code{TRUE}. If
asreml is used fitting spatial models is optional.}

\item{design}{A string specifying the experimental design. Either "ibd"
(incomplete block design), "res.ibd" (resolvable incomplete block design),
"rcbd" (randomized complete block design), "rowcol" (row column design) or
"res.rowcol" (resolvable row column design).}

\item{...}{Further arguments to be passed to \code{SpATS}, \code{lme4} or
\code{asreml}.}
}
\description{
Fit Single Trial Model using lme4
}
\seealso{
\code{\link{STRunModel}}
}
\keyword{internal}
