% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STModSpATS.R
\name{STModSpATS}
\alias{STModSpATS}
\title{Fit Single Trial Model using SpATS}
\usage{
STModSpATS(TD, traits, what = c("fixed", "random"), covariates = NULL,
  useCheckId = FALSE, trySpatial = FALSE, design = "rowcol",
  control = NULL, ...)
}
\arguments{
\item{TD}{An object of class \code{\link{TD}}.}

\item{traits}{A character vector specifying the selected traits.}

\item{what}{A character vector specififying whether "genotype" should
be fitted as "fixed" or "random". If not specified both models
are fitted.}

\item{covariates}{A string specifying (a) covariate name(s).}

\item{useCheckId}{Should a checkId be used as a fixed parameter in the model?\cr
If \code{TRUE} \code{TD} has to contain a column 'checkId'.}

\item{trySpatial}{Should spatial models be tried? Spatials models are can
only be fitted with SpATS and asreml.}

\item{design}{A string specifying the experimental design. One of "ibd"
(incomplete block design), "res.ibd" (resolvable incomplete block design),
"rcbd" (randomized complete block design), "rowcol" (row column design) or
"res.rowcol" (resolvable row column design).}

\item{control}{An optional list with control parameters to be passed to the
actual fitting funcions. For now only nSeg and nestDiv are valid and pass a
value to nseg and nest.div in \code{\link[SpATS]{PSANOVA}} respectively.}

\item{...}{Further arguments to be passed to \code{SpATS}, \code{lme4} or
\code{asreml}.}
}
\description{
Fit Single Trial Model using SpATS
}
\examples{
## Load data.
data(TDHeat05)

## Fit basic spatial model - no blocking, no replicates.
modSpATS1 <- STModSpATS(TD = TDHeat05, trait = "yield")

## Fit spatial model including replicates - no blocking.
modSpATS2 <- STModSpATS(TD = TDHeat05, trait = "yield", design = "res.rowcol")

## Fit spatial model including replicates and blocking.
modSpATS3 <- STModSpATS(TD = TDHeat05, trait = "yield", design = "res.ibd")

}
\seealso{
\code{\link{STRunModel}}
}
