% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTD.R
\name{[.TD}
\alias{[.TD}
\title{Function for extracting for objects of class TD that keeps attributes
including class.}
\usage{
\method{[}{TD}(x, i, ...)
}
\description{
Function for extracting for objects of class TD that keeps attributes
including class.
}
\keyword{internal}
