% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSSA.R
\name{summary.SSA}
\alias{summary.SSA}
\title{Summarizing objects of class \code{SSA}}
\usage{
\method{summary}{SSA}(object, trait = NULL, digits = max(getOption("digits")
  - 2, 3), nBest = 20, sortBy = if (!is.null(object$mFix)) "BLUEs" else
  "BLUPs", naLast = TRUE, decreasing = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{SSA}.}

\item{trait}{a string indicating the trait to summarize over. If
\code{trait = NULL} and only one trait is modelled this trait is summarized.}

\item{digits}{an integer. The number of significant digits to use when
printing.}

\item{nBest}{an integer. The number of the best genotypes (sorted by either
BLUEs or BLUPs) to print. If \code{NA} all genotypes are printed.}

\item{sortBy}{a character string specifying how the genotypes will be sorted.
The options are \code{"BLUEs"}, \code{"BLUPs"} and \code{NA} (i.e. no sort).}

\item{naLast}{Should missing values in the data be put last when sorting?}

\item{decreasing}{should the sort order be decreasing?}

\item{...}{further arguments passed to \code{\link[stats]{printCoefmat}}.}
}
\description{
\code{summary} method for class \code{SSA}.
}
\examples{
# Run a single trait analysis using SpATS.
myModel <- STRunModel(TD = TDHeat05, design = "res.rowcol", traits = "yield")
# Print a summary of the model.
summary(myModel)

}
