% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTD.R
\name{summary.TD}
\alias{summary.TD}
\title{Summarizing objects of class \code{TD}}
\usage{
\method{summary}{TD}(object, ..., traits, what = c("nObs", "nMiss", "mean",
  "median", "min", "max", "lowerQ", "upperQ", "var"))
}
\arguments{
\item{object}{An object of class TD.}

\item{...}{Further arguments - currently not used.}

\item{traits}{A character vector specifying the name(s) of the traits
to be summarised.}

\item{what}{A character vector indicating which statistics should be computed.\cr
If \code{what = "all"} all available statistics are computed.\cr
Possible options are\cr
\describe{
\item{nVals}{the number of values}
\item{nObs}{the number of observations}
\item{nMiss}{the number of missing values}
\item{mean}{the mean}
\item{median}{the median}
\item{min}{the minimum}
\item{max}{the maximum}
\item{range}{the range (maximum - minimum)}
\item{lowerQ}{the lower (25\%) quantile}
\item{upperQ}{the upper (75\%) quantile}
\item{sd}{the standard deviation}
\item{seMean}{standard error of mean}
\item{var}{the variance}
\item{seVar}{the standard error of variance}
\item{CV}{the coefficient of variation}
\item{sum}{the sum}
\item{sumSq}{sum of squares}
\item{uncorSumSq}{uncorrected sum of squares}
\item{skew}{the skewness}
\item{seSkew}{the standard error of skewness}
\item{kurt}{the kurtosis}
\item{seKurt}{the standard error of kurtosis}
}}
}
\value{
A table containing the selected summary statistics.
}
\description{
\code{summary} method for class \code{TD}.
}
\examples{
summary(object = TDHeat05, traits = "yield")

}
\seealso{
\code{\link{createTD}}
}
