% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{supprWarn}
\alias{supprWarn}
\title{Helper function for suppressing a single warning message.}
\usage{
supprWarn(expression, message)
}
\description{
Helper function for suppressing a single warning message.
}
\keyword{internal}
