% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDHeat05.R
\docType{data}
\name{TDMaize}
\alias{TDMaize}
\title{Test data maize}
\format{An object of class \code{TD} (inherits from \code{data.frame}) with 1688 rows and 3 columns.}
\usage{
TDMaize
}
\description{
Test data maize
}
\keyword{datasets}
