% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test4way}
\alias{test4way}
\title{Fake data of 4way cross for unit testing.}
\format{An object of class \code{4way} (inherits from \code{cross}) of length 2.}
\usage{
test4way
}
\description{
Base data taken from qtl package. Decreased size by removing most individuals
and SNPs. To be used for (unit) testing only.
}
\keyword{internal}
