% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDHeat05.R
\docType{data}
\name{testBc}
\alias{testBc}
\title{Fake data of backcross for unit testing.}
\format{An object of class \code{bc} (inherits from \code{cross}) of length 2.}
\usage{
testBc
}
\description{
Base data taken from qtl package. Decreased size by removing most individuals
and SNPs. To be used for (unit) testing only.
}
\keyword{internal}
