% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDHeat05.R
\docType{data}
\name{testData}
\alias{testData}
\title{Random test data for unit testing.}
\format{An object of class \code{data.frame} with 90 rows and 12 columns.}
\usage{
testData
}
\description{
Random test data for unit testing.
}
\keyword{datasets}
