% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDHeat05.R
\docType{data}
\name{testF2}
\alias{testF2}
\title{Fake data of f2 cross for unit testing.}
\format{An object of class \code{f2} (inherits from \code{cross}) of length 3.}
\usage{
testF2
}
\description{
Fake data of f2 cross for unit testing.
}
\keyword{datasets}
