% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tryCatchExt}
\alias{tryCatchExt}
\title{Custom tryCatch to return result, errors and warnings.
Copied from http://stackoverflow.com/a/24569739/2271856.}
\usage{
tryCatchExt(expr)
}
\description{
Custom tryCatch to return result, errors and warnings.
Copied from http://stackoverflow.com/a/24569739/2271856.
}
\keyword{internal}
