% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDHeat05.R
\docType{data}
\name{wheatAus}
\alias{wheatAus}
\title{Data used in vignette}
\format{An object of class \code{data.frame} with 4992 rows and 8 columns.}
\usage{
wheatAus
}
\description{
Data used in vignette
}
\keyword{datasets}
