% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDHeat05.R
\docType{data}
\name{wheatChl}
\alias{wheatChl}
\title{Data used in vignette}
\format{An object of class \code{data.frame} with 4000 rows and 13 columns.}
\usage{
wheatChl
}
\description{
Data used in vignette
}
\keyword{datasets}
