% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrnToErr}
\alias{wrnToErr}
\title{Helper function for converting certain asreml warnings to errors.}
\usage{
wrnToErr(model)
}
\description{
Helper function for converting certain asreml warnings to errors.
}
\keyword{internal}
