% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAMMI.R
\name{AMMI}
\alias{AMMI}
\alias{createAMMI}
\title{S3 class AMMI}
\usage{
createAMMI(envScores, genoScores, importance, anova, fitted, trait, envMean,
  genoMean, overallMean)
}
\arguments{
\item{envScores}{a matrix containing environment scores}

\item{genoScores}{a matrix containing genotypic scores}

\item{importance}{a data.frame containing the importance of the principal components}

\item{anova}{a data.frame containing anova scores of the AMMI analysis}

\item{fitted}{a matrix containing fitted values from the AMMI model}

\item{trait}{a character value indicating the analysed trait}

\item{envMean}{a numerical vector containing the means per environment}

\item{genoMean}{a numerical vector containing the means per genotype}

\item{overallMean}{a numerical value containing the overall mean}
}
\description{
Function for creating objects of S3 class AMMI.\cr
\code{\link{print}}, \code{\link{summary}}, \code{\link{plot}} and
\code{\link{report}} methods are available.
}
\seealso{
\code{\link{plot.AMMI}}, \code{\link{report.AMMI}}
}
\author{
Bart-Jan van Rossum
}
