% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAMMI.R
\name{fitted.AMMI}
\alias{fitted.AMMI}
\title{Extract fitted values.}
\usage{
\method{fitted}{AMMI}(object, ...)
}
\arguments{
\item{object}{An object of class AMMI}

\item{...}{Not used.}
}
\description{
Extract the fitted values for an object of class AMMI.
}
