% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFW.R
\name{fitted.FW}
\alias{fitted.FW}
\title{Extract fitted values.}
\usage{
\method{fitted}{FW}(object, ...)
}
\arguments{
\item{object}{An object of class FW}

\item{...}{Not used.}
}
\description{
Extract the fitted values for an object of class FW.
}
