% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeAmmi.R
\name{gxeAmmi}
\alias{gxeAmmi}
\title{AMMI analysis}
\usage{
gxeAmmi(TD, trait, nPC = 2, center = TRUE, scale = FALSE)
}
\arguments{
\item{TD}{an object of class \code{\link{TD}}.}

\item{trait}{A character string specifying the trait to be analyzed.}

\item{nPC}{An integer specifying the number of principal components used
as multiplicative term of genotype-by-environment interaction.}

\item{center}{Should the variables be shifted to be zero centered?}

\item{scale}{Should the variables be scaled to have unit variance before
the analysis takes place?}
}
\value{
an object of class \code{\link{AMMI}}, a list containing
\item{envScores}{a matrix with environmental scores.}
\item{genoScores}{a matrix with genotypic scores.}
\item{importance}{a data.frame containing the importance of the principal
components.}
\item{anova}{a data.frame containing anova scores of the AMMI analysis.}
\item{fitted}{a matrix containing fitted values from the AMMI model.}
\item{trait}{a character vector containing the analyzed trait.}
\item{envMean}{a numerical vector containing the means per environment.}
\item{genoMean}{a numerical vector containing the means per genotype.}
\item{overallMean}{a numerical value containing the overall mean.}
}
\description{
This function fits a model which involves the Additive Main effects (i.e.
genotype and environment) along with the Multiplicative Interaction effects
of principal component analysis (PCA).
}
\examples{
## Run AMMI
geAmmi <- gxeAmmi(TD = TDMaize, trait = "yld")
## Create report
report(geAmmi, outfile = "./testReports/reportAmmi.pdf")

}
