% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeFw.R
\name{gxeFw}
\alias{gxeFw}
\title{Finlay-Wilkinson analysis}
\usage{
gxeFw(TD, trait, maxIter = 15, tol = 0.001, sorted = c("ascending",
  "descending", "none"))
}
\arguments{
\item{TD}{an object of class \code{\link{TD}}.}

\item{trait}{A character string specifying the trait to be analyzed.}

\item{maxIter}{An integer specifying the maximum number of iterations in
the algorithm.}

\item{tol}{A positive numerical value specifying convergence tolerance of the
algorithm.}

\item{sorted}{A character string specifying the sorting order of the
estimated values in the output.}
}
\value{
An object of class \code{\link{FW}}, a list containing:
\item{estimates}{A data.frame containing the estimated values.}
\item{anova}{A data.frame containing anova scores of the FW analysis.}
\item{envEffs}{A data.frame containing the environmental effects.}
\item{TD}{The object of class TD on which the analysis was performed.}
\item{fittedGeno}{A numerical vector containing the fitted values for the
genotypes.}
\item{trait}{A character string containing the analyzed trait.}
\item{nGeno}{A numerical value containing the number of genotypes in the
analysis.}
\item{nEnv}{A numerical value containing the number of environments in the
analysis.}
\item{tol}{A numerical value containing the tolerance used during the
analysis.}
\item{iter}{A numerical value containing the number of iterations for the
analysis to converge.}
}
\description{
This function performs a Finlay-Wilkinson analysis of data classified by two
factors.
}
\examples{
## Run Finlay-Wilkinson analysis on TDMaize.
geFW <- gxeFw(TDMaize, trait = "yld")
## Summarize results.
summary(geFW)
## Create a scatterplot of the results.
plot(geFW, plotType = "scatter")
\dontrun{
## Create a report summarizing the results.
report(geFW, outfile = "./testReports/reportFW.pdf")
}

}
\references{
Finlay, K.W. & Wilkinson, G.N. (1963). The analysis of adaptation
in a plant-breeding programme. Australian Journal of Agricultural
Research, 14, 742-754.
}
\seealso{
\code{\link{FW}}, \code{\link{plot.FW}}, \code{\link{report.FW}}
}
