% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeMegaEnv.R
\name{gxeMegaEnv}
\alias{gxeMegaEnv}
\title{Form mega-environments based on winning genotypes from an AMMI model}
\usage{
gxeMegaEnv(TD, trait, method = c("max", "min"), sumTab = TRUE)
}
\arguments{
\item{TD}{an object of class \code{\link{TD}}.}

\item{trait}{A character string specifying the trait to be analyzed.}

\item{method}{A character string indicating the criterion to determine
the best genotype per environment, either \code{"max"} or \code{"min"}.}

\item{sumTab}{Should a summary table be added as an attribute to
the output and be printed?}
}
\value{
The input object of class \code{\link{TD}} with an added extra
column megaEnv.
}
\description{
This function fits an AMMI model and then using the fitted values produces
a new factor based on the winning genotype in each environment. This factor
is added as a column megaEnv to the input data. If a column megaEnv already
exists the existing data is overwritten with a warning.
}
\examples{
## Calculate mega-environments for TDMaize and print a summary of the results.
TDmegaEnv <- gxeMegaEnv(TD = TDMaize, trait = "yld")
## Calculate new mega-environments based on the genotypes with the lowest
## value per environment.
TDmegaEnv2 <- gxeMegaEnv(TD = TDmegaEnv, trait = "yld", method = "min")

}
