% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeTable.R
\name{gxeTable}
\alias{gxeTable}
\title{Compute BLUPS based on a set of mega environments}
\usage{
gxeTable(
  TD,
  trials = names(TD),
  trait,
  useYear = FALSE,
  engine = c("lme4", "asreml"),
  ...
)
}
\arguments{
\item{TD}{An object of class \code{\link{TD}}.}

\item{trials}{A character string specifying the trials to be analyzed. If
not supplied, all trials are used in the analysis.}

\item{trait}{A character string specifying the trait to be analyzed.}

\item{useYear}{Should year be used for modeling (as years within
trials). If \code{TRUE}, \code{TD} should contain a column "year".}

\item{engine}{A character string specifying the engine used for modeling.
Either "lme4" or "asreml".}

\item{...}{Further parameters passed to either \code{asreml} or \code{lmer}.}
}
\value{
A list consisting of two data.frames, \code{predictedValue}
containing BLUPs per genotype per mega environment and \code{standardError}
containing standard errors for those BLUPs.
}
\description{
This function calculates Best Lineair Unbiased Predictors (BLUPS) and
associated standard errors based on a set of mega environments.
}
\examples{
## Compute mega environments for TDMaize.
TDMegaEnv <- gxeMegaEnv(TD = TDMaize, trait = "yld", sumTab = FALSE)

## Compute BLUPS and standard errors for those mega environments.
geTab <- gxeTable(TD = TDMegaEnv, trait = "yld")

}
