% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeVarComp.R
\name{gxeVarComp}
\alias{gxeVarComp}
\title{Selects the best variance-covariance model for a set of environments}
\usage{
gxeVarComp(TD, trait, engine = "asreml", criterion = "BIC", ...)
}
\arguments{
\item{TD}{an object of class \code{\link{TD}}.}

\item{trait}{A character string specifying the trait to be analyzed.}

\item{engine}{A string specifying the name of a mixed modelling engine to be used.}

\item{criterion}{A string specifying a goodness-of-fit criterion, i.e., "AIC" or "BIC".}

\item{...}{Further arguments to be passed to \code{asreml}.}
}
\value{
A list object consisting of the fitted model objects, a string specifying
the best model and its related goodness-of-fit criterion.
}
\description{
This function selects the best covariance structure for genetic correlations
between environments. It fits a range of variance-covariance models to
compare (identity, compound symmetry, diagonal, heterogeneous compound
symmetry, first order factor analysis, second order factor analysis,
unstructured), and selects the best one using a goodness-of-fit criterion.
}
\note{
If \code{engine = "lme4"}, only the compound symmetry model can be fitted.
}
\examples{
## Fit models.
geVarComp1 <- gxeVarComp(TD = TDMaize, trait = "yld", engine = "lme4",
                        criterion = "BIC")
## Display results.
summary(geVarComp1)

\dontrun{
geVarComp2 <- gxeVarComp(TD = TDMaize, trait = "yld", engine = "asreml",
                        criterion = "BIC")
report(geVarComp2, outfile = "./testReports/reportVarComp.pdf")
}

}
