% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarComp.R
\name{herit}
\alias{herit}
\title{Compute heritability}
\usage{
herit(varComp)
}
\arguments{
\item{varComp}{An object of class varComp.}
}
\description{
Compute the heritability based on the fitted model.
}
