% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiQTLFit.R
\name{multiQTLFit}
\alias{multiQTLFit}
\title{Fit an additive multi QTL model}
\usage{
multiQTLFit(QTLDet, selection = c("backward", "none"), thr = 0.05, ...)
}
\arguments{
\item{QTLDet}{An object of class \code{\link{QTLDet}}.}

\item{selection}{A character string indicating whether backward selection
should be applied or no selection at all.}

\item{thr}{A positive numerical value indicating the threshold for dropping
terms in the backwards elemination process.}

\item{...}{Further parameters to be passed on to underlying functions used
for qtl detection, \code{\link[qtl]{scanone}} when \code{type} is "MR"
or "SIM" and \code{\link[qtl]{cim}} when \code{type} is "CIM".}
}
\value{
An object of class \code{\link{multiQTL}}, a list containing:
\item{scores}{A data.frame containing the lod scores.}
\item{peaks}{A data.frame containing the peaks found.}
\item{type}{A character string indicating the type of QTLDetection performed.}
\item{cross}{An object of class cross in the \code{qtl} package on which
the analysis has been performed.}
\item{trait}{A character string indicating the trait for which the analysis
is done.}
\item{info}{A list containing information on the settings used for
QTL Detection, i.e. step, threshold and window.}
}
\description{
An additive multi QTL model is fitted based on the peaks in the QTLDet object.
Fitting is done using the \code{\link[qtl]{fitqtl}} function in the qtl package.
After fitting the model backward elemination is done until all markers in
the model have a significant P-value.
}
\examples{
## Read the data
F2 <- qtl::read.cross(format="csv",
                      file = system.file("extdata", "F2maize_geno.csv",
                                        package = "RAP"),
                      genotypes = c("AA", "AB", "BB"),
                      alleles = c("A", "B"), estimate.map = FALSE)
## Perform QTL detection using simple interval mapping.
QTLDet <- QTLDetect(cross = F2, trait = "trait", type = "SIM")
## Fit a multi QTL model.
multiFit <- multiQTLFit(QTLDet)
## Summarize results.
summary(multiFit)
\dontrun{
## Create a pdf report summarizing results.
report(multiFit, outfile = "./testReports/reportMultiQTLFit.pdf")
}

}
\references{
Broman et al. (2003) R/qtl: QTL mapping in experimental crosses.
Bioinformatics 19:889-890
}
\seealso{
\code{\link[qtl]{fitqtl}}
}
