% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlierSSA}
\alias{outlierSSA}
\title{Identifying outliers in objects of class SSA}
\usage{
outlierSSA(SSA, trial = NULL, traits, rLimit = NULL,
  commonFactors = NULL, verbose = TRUE)
}
\arguments{
\item{SSA}{An object of class \code{\link{SSA}}.}

\item{trial}{A character string specifying the trial for which outliers
should be identified. If \code{NULL} and \code{SSA} contains only one trial
that trial is used.}

\item{traits}{A character vector specifying the names of the traits for
which outliers should be identified.}

\item{rLimit}{A numerical value used for determining when a value is
considered an outlier. All observations with standardized residuals
exceeding \code{rLimit} will be marked as outliers.}

\item{commonFactors}{A character vector specifying the names of columns
in \code{TD} used for selecting observations that are similar to the
outliers. If \code{commonFactors = NULL} only outliers are reported and
no similar observations.}

\item{verbose}{Should the outliers be printed to the console?}
}
\value{
A list with two components:
\itemize{
\item{indicator - a data.frame containing logical values indicating if the
observation is an outlier.}
\item{outliers - a data.frame containing the outliers and observations
similar to the outliers as defined by \code{commonFactors}}
}
}
\description{
Function to identify observations with standardized residuals exceeding
\code{rLimit}. If not provided \code{rLimit} is computed as
\code{qnorm(1 - 0.5 / rDf)} where \code{rDf} are the residual degrees
of freedom for the model. This value is then restricted to the interval
2..4. Alternatively a custom limit may be provided.\cr
A summary is printed of outliers and observations that have the same
value for \code{commonFactors}. The latter ones will be marked as
similar to distinguish them from the former ones.
}
\examples{
## Fit a model using lme4.
myModel <- STRunModel(TD = TDHeat05, traits = "yield" ,
                      design = "res.rowcol", engine = "lme4")
## Detect outliers in the standardized residuals of the fitted model.
outliers <- outlierSSA(SSA = myModel, traits = "yield")

}
