% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFW.R
\name{plot.FW}
\alias{plot.FW}
\title{Plot function for class FW}
\usage{
\method{plot}{FW}(
  x,
  ...,
  plotType = c("scatter", "line", "trellis", "scatterFit"),
  order = c("ascending", "descending"),
  lineTrait = c("fitted", "raw"),
  genotypes = NULL,
  output = TRUE
)
}
\arguments{
\item{x}{An object of class FW.}

\item{...}{Further graphical parameters passed on to actual plot function.}

\item{plotType}{A character string indicating which plot should be made.
Either "scatter", "line" or "trellis" for creating a scatter
plot of genotypic means, mse and sensitivities, a plot of fitted lines for
each genotype or a trellis plot of the individual genotype slopes
respectively.}

\item{order}{A character string specifying whether the results in the line
plot should be ordered in an increasing (or decreasing) order of
sensitivities. Ignored if \code{plotType} is not "line".}

\item{lineTrait}{A character string specifying whether in the line plot the
"fitted" or the "raw" data should be plotted. Ignored if \code{plotType} is
not "line".}

\item{genotypes}{An optional character string containing the genotypes to
be plotted in the trellis plot. If \code{NULL} all genotypes are plotted.
If more than 64 genotypes are selected, only the first 64 are plotted.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE}, only a list of ggplot objects is invisibly returned.}
}
\value{
A plot depending on \code{plotType}.
}
\description{
Four types of plot can be made. A scatter plot for genotypic mean,
mean squared deviation and sensitivity, a line plot with fitted lines for
each genotype, a trellis plot with individual slopes per genotype and a
scatter plot of fitted values in the worst and best trial.\cr
It is possible to select genotypes for the trellis plot using the
\code{genotypes} parameter. If there are more than 64 genotypes, only the
first 64 are plotted in the trellis plot.
}
\examples{
## Run Finlay-Wilkinson analysis.
geFW <- gxeFw(TD = TDMaize, trait = "yld")

## Create a scatter plot.
plot(geFW)

## Create a line plot.
plot(geFW, plotType = "line")

## Create a line plot showing raw data value for genotypes and fitted lines.
## Display trials in descending order.
plot(geFW, plotType = "line", order = "descending", lineTrait = "raw")

## Create a trellis plot.
plot(geFW, plotType = "trellis")

## Create a scatter plot of fitted values for the worst and best trials.
plot(geFW, plotType = "scatterFit")

}
