% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultiQTL.R
\name{plot.multiQTL}
\alias{plot.multiQTL}
\title{Plot function for class multiQTL}
\usage{
\method{plot}{multiQTL}(x, ...,
  title = "QTL estimates and confidence intervals", output = TRUE)
}
\arguments{
\item{x}{An object of class multiQTL}

\item{...}{Further graphical parameters. Currently not used.}

\item{title}{A character string used as overall title for the plot.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned.}
}
\description{
Plotting function for objects of class multiQTL. Plots the estimates of the
QTLs in the final model with their respective confidence intervals.
}
\examples{
## Read the data
F2 <- qtl::read.cross(format="csv",
                      file = system.file("extdata",
                                        "F2_maize_practical3_ex2.csv",
                                        package = "RAP"),
                      genotypes = c("AA", "AB", "BB"),
                      alleles = c("A", "B"), estimate.map = FALSE)
## Perform QTL detection using simple interval mapping.
QTLDet <- QTLDetect(cross = F2, trait = "trait", type = "SIM")
## Fit a multi QTL model.
multiFit <- multiQTLFit(QTLDet)
## Plot the results.
plot(multiFit)

}
