% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLDet.R
\name{plot.QTLDet}
\alias{plot.QTLDet}
\title{Plot function for class QTLDet}
\usage{
\method{plot}{QTLDet}(x, ..., yLim = NA, title = x$trait,
  output = TRUE)
}
\arguments{
\item{x}{An object of class QTLDet}

\item{...}{Not used}

\item{yLim}{A numerical value for the upper limit of the y-axis. If
\code{NA} the limit is determined based on the data.}

\item{title}{A character value for the title of the plot. If not supplied
the trait used in the qTL Detection is used as plot title.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned.}
}
\description{
Function for creating a manhattan plot for objects of class QTLDet.
}
