% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLDet.R
\name{plot.QTLDet}
\alias{plot.QTLDet}
\title{Plot function for class QTLDet}
\usage{
\method{plot}{QTLDet}(x, ...)
}
\arguments{
\item{x}{an object of class QTLDet}

\item{...}{not unused}
}
\description{
Function for creating a manhattan plot for objects of class QTLDet.
}
