% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSSA.R
\name{plot.SSA}
\alias{plot.SSA}
\title{Plot Function for Class SSA}
\usage{
\method{plot}{SSA}(x, ..., trait = NULL, what = ifelse(is.null(x$mFix),
  "random", "fixed"), plotType = c("base", "spatial"))
}
\arguments{
\item{x}{An object of class SSA.}

\item{...}{Other graphical parameters (see \code{\link[lattice]{xyplot}}
for details).}

\item{trait}{a string indicating the trait to summarize over. If
\code{trait = NULL} and only one trait is modelled this trait is summarized.}

\item{what}{A character string indicating whether the model with
genotype fixed or random should be plotted. If \code{x} contains only
one model this model is chosen automatically.}

\item{plotType}{character string indicating whether \code{base} plots or
\code{spatial} plots should be made.}
}
\description{
This function draws either four base plots
\itemize{
\item{a histogram of the residuals}
\item{a normal Q-Q plot}
\item{a residuals vs fitted values plot}
\item{an absolute residuals vs fitted values plot}
}
or five or (in case SpATS is used for modelling) six spatial plots
\itemize{
\item{a spatial plot of the raw data}
\item{a spatial plot of the fitted data}
\item{a spatial plot of the residuals}
\item{a spatial plot of the estimated spatial trend (SpATS only)}
\item{a spatial plot of the BLUEs or BLUPs}
\item{a histogram of the BLUEs or BLUPs}
}
Spatial plots can only be made if the data contains both row and column information.
}
\examples{
# Run a single trait analysis using SpATS.
myModel <- STRunModel(TD = TDHeat05, design = "res.rowcol", traits = "yield")
# Create base plots
plot(myModel, what = "fixed", plotType = "base")
# Create spatial plots
plot(myModel, what = "fixed", plotType = "spatial")

}
\seealso{
\code{\link{SSA}}
}
