% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSSA.R
\name{plot.SSA}
\alias{plot.SSA}
\title{Plot function for class SSA}
\usage{
\method{plot}{SSA}(x, ..., trial = NULL, trait = NULL, what = NULL,
  plotType = c("base", "spatial"))
}
\arguments{
\item{x}{An object of class SSA.}

\item{...}{Further graphical parameters (see \code{\link[lattice]{xyplot}}
for details).}

\item{trial}{a character string indicating the trial to plot. If
\code{trial = NULL} and only one trial is modelled this trial is plotted.}

\item{trait}{a character string indicating the trait to plot. If
\code{trait = NULL} and only one trait is modelled this trait is plotted.}

\item{what}{A character string indicating whether the fitted model with
genotype as fixed or genotype as random factor should be plotted.
If \code{x} contains only one model this model is chosen automatically.}

\item{plotType}{character string indicating whether \code{base} plots or
\code{spatial} plots should be made.}
}
\description{
This function draws either four base plots:
\itemize{
\item{A histogram of the residuals}
\item{A normal Q-Q plot}
\item{A residuals vs fitted values plot}
\item{An absolute residuals vs fitted values plot}
}
or five or (in case SpATS is used for modelling) six spatial plots:
\itemize{
\item{A spatial plot of the raw data}
\item{A spatial plot of the fitted data}
\item{A spatial plot of the residuals}
\item{A spatial plot of the estimated spatial trend (SpATS only)}
\item{A spatial plot of the BLUEs or BLUPs}
\item{A histogram of the BLUEs or BLUPs}
}
Spatial plots can only be made if the data contains both row and column
information.
}
\examples{
## Run a single trait analysis using SpATS.
myModel <- STRunModel(TD = TDHeat05, design = "res.rowcol", traits = "yield")
## Create base plots.
plot(myModel, what = "fixed", plotType = "base")
## Create spatial plots.
plot(myModel, what = "fixed", plotType = "spatial")

}
\seealso{
\code{\link{SSA}}
}
