% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStability.R
\name{plot.stability}
\alias{plot.stability}
\title{Plot function for class stability}
\usage{
\method{plot}{stability}(x, ..., colorBy = NULL, output = TRUE)
}
\arguments{
\item{x}{An object of class stability.}

\item{...}{Further arguments - not used.}

\item{colorBy}{A character string indicating a column in the \code{TD} used
as input for the stability analysis by which the genotypes should be
colored. If \code{NULL} all genotypes will be colored black.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned.}
}
\value{
Plots of stability measures against means.
}
\description{
Function for creating scatter plots of computed stability measures against
the means.
}
\examples{
## Compute three stability measures for TDMaize.
geStab <- gxeStability(TD = TDMaize, trait = "yld")
## Create scatter plots of the computed stability measures against the means.
plot(geStab)

}
