% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTD.R
\name{plot.TD}
\alias{plot.TD}
\title{Plot Function for Class TD}
\usage{
\method{plot}{TD}(x, ..., trait, plotType = c("box", "hist", "scatter",
  "cor"))
}
\arguments{
\item{x}{An object of class TD.}

\item{...}{Other graphical parameters. For boxplots, histograms and
scatterplots all graphical parameters can be changed, for the correlation
only \code{"main"} will be used and other parameters are ignored.}

\item{trait}{A character string specifying the name of the traits to be plotted.}

\item{plotType}{A character string indicating which plot should be made.
Either \code{"box"} for a boxplot, \code{"hist"} for histograms,
\code{"scatter"} for scatter plots and correlations or \code{"cor"} for a
plot of the correlations between environments.}
}
\value{
One of four plots depending on \code{plotType}.
}
\description{
Four types of plot can be made. Boxplots and histograms can be made for
all objects of class \code{\link{TD}}. In case there is a column \code{"env"}
in TD boxplots and histograms will be made per environment.\cr
Scatterplots and plots of correlation between environments can only be made
if a column \code{"env"} is present in TD and will result in an error if this
is not the case.
}
\examples{
plot(TDMaize, trait = "yld")

}
\seealso{
\code{\link{TD}}
}
