% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTD.R
\name{plot.TD}
\alias{plot.TD}
\title{Plot function for class TD}
\usage{
\method{plot}{TD}(x, ..., plotType = c("layout", "map", "box"),
  trials = names(x), traits = NULL, output = TRUE)
}
\arguments{
\item{x}{An object of class TD.}

\item{...}{Not currently used.}

\item{plotType}{A character string indicating which plot should be made.
This can be "layout" for a plot of the field layout for the diffent trials
in the TD object. This is only possible if the data contains row and column
information. Alternatively, for \code{plotType = "map"} a plot will be made
depicting the trials on a country map. This is only possible if lattitude
and longitude of the trials are available.}

\item{trials}{A character vector indicating the trials to be plotted when
plotting field layouts. Only used if \code{plotType = "layout"}}

\item{traits}{A character vector indicating the traits to be plotted in
a boxplot. Only used if \code{plotType = "box"}}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned.}
}
\description{
Plotting function for objects of class TD. Plots either the layout of the
different trials within the TD object or locates the trials on a map.
Mapping the trials is done based on lattitude and longitude that can be
added when creating an object of class TD. The countries in which the trials
are located will be plotted on a single map and the location of the trials
will be indicated on this map.
}
