% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarComp.R
\name{plot.varComp}
\alias{plot.varComp}
\title{Plot function for class varComp}
\usage{
\method{plot}{varComp}(x, ..., output = TRUE)
}
\arguments{
\item{x}{An object of class varComp}

\item{...}{Not used}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned.}
}
\description{
Plot function for class varComp.
}
