% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarComp.R
\name{plot.varComp}
\alias{plot.varComp}
\title{Plot Function for Class varComp}
\usage{
\method{plot}{varComp}(x, ...)
}
\arguments{
\item{x}{An object of class varComp}

\item{...}{not unused}
}
\description{
Function for plotting a heatmap of the correlation matrix for objects of
class varComp.
}
\examples{
\dontrun{
geVarComp <- gxeVarComp(TD = TDMaize, trait = "yld", engine = "asreml")
plot(geVarComp)
}

}
