% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTLDetect.R
\name{QTLDetect}
\alias{QTLDetect}
\title{QTL detection}
\usage{
QTLDetect(cross, trait, type = c("MR", "SIM", "CIM"), step = 5,
  thrType = c("liji", "bonferroni", "fixed"), thrAlpha = 0.05,
  thrDist = 4, thrFixed = 3, window = 30, ...)
}
\arguments{
\item{cross}{An object of class cross created by the qtl package.}

\item{trait}{A character string indicating the trait to be analyzed.}

\item{type}{A character string indicating the type of QTL detection to be
performed. Either "MR" (Marker Response), "SIM" (Simple Interval Mapping)
or "CIM" (Composite Interval Mapping)}

\item{step}{A numerical value indicating the maximum distance (in cM)
between positions at which the genotype probabilities are calculated. If
\code{step = 0} probabilities are only calculated at the marker locations.}

\item{thrType}{A character string indicating the algorithm to calculate the
lower threshold for the lodscore of the peaks. See details.}

\item{thrAlpha}{A numerical value between 0 and 1 used as enumerator in
calculting the modified Bonferroni correction.}

\item{thrDist}{A positive numerical value indicating the assumed fixed
distance on the genome for independent tests (in Centimorgan). Only used
when \code{thrType = "Bonferroni"}}

\item{thrFixed}{A numerical value indicating a lower threshold for the
lodscore of the peaks. Only used when \code{thrType = "fixed"}}

\item{window}{A numerical value indicating the window (in cM) used when
selecting peaks.}

\item{...}{Further parameters to be passed on to underlying functions used
for qtl detection, \code{\link[qtl]{scanone}} when \code{type} is "MR"
or "SIM" and \code{\link[qtl]{cim}} when \code{type} is "CIM".}
}
\value{
An object of class \code{\link{QTLDet}}, a list containing:
\item{scores}{A data.frame containing the lod scores.}
\item{peaks}{A data.frame containing the peaks found.}
\item{type}{A character string indicating the type of QTL detection
performed.}
\item{cross}{The object of class cross used for QTL detection.}
\item{trait}{A character string indicating the trait for which the detection
was done.}
\item{info}{A list containing information on the settings used for
QTL detection, i.e. step, threshold and window.}
}
\description{
This function is essentially a wrapper for \code{\link[qtl]{scanone}} and
\code{\link[qtl]{cim}} in the qtl package. Depending on \code{type} one
of these functions is used for QTL detection. After this is done, from
the set of candidate QTLs that are returned proper peaks are selected
by an iterative process using the \code{thr}(eshold) and \code{window}
provided. All resulting peaks will have a LOD score above \code{thr}
and the distance between pairs of peaks will always be at least
the value given as \code{window}.
}
\details{
The option \code{thrType} specifies the method for calculating the number of
tests to be used as the denominator in a modified Bonferroni correction. The
enumeration is given by \code{thrAlpha}.
\code{liji} uses the effective number of independent tests, as described by
Li & Ji (2005). \code{"bonferroni"} assumes one independent test at every
fixed distance on the genome, defined by the \code{thrDist} (default 4
centiMorgans). The threshold is expressed as the P-value on a -log10 scale.
By setting {thrType = "fixed"} all the above is ignored and a fixed value is
used as threshold instead. This value can be supplied in \code{thrFixed}
}
\note{
Composite Interval Mapping is not implemented for 4 way crosses.
}
\examples{
## Read the data.
F2 <- qtl::read.cross(format="csv",
                      file = system.file("extdata",
                                        "F2_maize_practical3_ex2.csv",
                                        package = "RAP"),
                      genotypes = c("AA", "AB", "BB"),
                      alleles = c("A", "B"), estimate.map = FALSE)
## Perform a composite interval mapping for detecting QTLs.
QTLDet <- QTLDetect(cross = F2, trait = "trait", type = "CIM")
## Summarize results.
summary(QTLDet)
## Create a manhattan plot of the results.
plot(QTLDet)
\dontrun{
## Create a pdf report summarizing the results.
report(QTLDet, outfile = "./testReports/reportQTLDectection.pdf")
}

## Perform a simple interval mapping for detecting QTLs.
## Choose custom step, threshold and window sizes.
QTLDet2 <- QTLDetect(cross = F2, trait = "trait", type = "SIM", step = 15,
                    thr = 2.5, window = 50)
summary(QTLDet2)

}
\references{
Broman et al. (2003) R/qtl: QTL mapping in experimental crosses.
Bioinformatics 19:889-890

Cheverud, J.M. (2001). A simple correction for multiple
comparisons in interval mapping genome scans. Heredity, 87, 52-58.

Li, J, & Ji, L. (2005). Adjusting multiple testing in multilocus
analyses using the eigenvalues of a correlation matrix. Heredity, 95,
221-227.
}
\seealso{
\code{\link[qtl]{scanone}}, \code{\link[qtl]{cim}}
}
