% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeVarComp.R
\name{qvInitial}
\alias{qvInitial}
\title{Helper function for computing initial values.}
\usage{
qvInitial(TD, trait, useWt = FALSE, vcmodel = c("identity", "cs",
  "diagonal", "hcs", "outside", "fa", "fa2", "unstructured"), fixed = NULL,
  ...)
}
\description{
Replicates '_qvInitial' procedure (S. J. Welham 15/05/09) in GenStat
}
\keyword{internal}
