% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{createTD}
\alias{fitTD}
\alias{SSAtoTD}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{statgenSSA}{\code{\link[statgenSSA]{createTD}}, \code{\link[statgenSSA]{fitTD}}, \code{\link[statgenSSA]{SSAtoTD}}}
}}

