% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultiQTL.R
\name{report.multiQTL}
\alias{report.multiQTL}
\title{Report method for class multiQTL}
\usage{
\method{report}{multiQTL}(x, ..., outfile = NULL)
}
\arguments{
\item{x}{an object of class multiQTL.}

\item{...}{further arguments passed on from other functions - not used yet.}

\item{outfile}{a character string, the name and location of the output .pdf and .tex
file for the report. If \code{NULL} a report will be created in the current working
directory.}
}
\description{
A pdf report will be created containing a summary of multiQTL analysis.
Simultaneously the same report will be created as a tex file.
}
