% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLDet.R
\name{report.QTLDet}
\alias{report.QTLDet}
\title{Report method for class QTLDet}
\usage{
\method{report}{QTLDet}(x, ..., outfile = NULL)
}
\arguments{
\item{x}{An object of class QTLDet.}

\item{...}{Further arguments passed on from other functions - not used yet.}

\item{outfile}{A character string, the name and location of the output .pdf
and .tex file for the report. If \code{NULL}, a report with a default name
will be created in the current working directory.}
}
\value{
A pdf and tex report.
}
\description{
A pdf report will be created containing a summary of a QTLDet analysis.
Simultaneously the same report will be created as a tex file.
}
\examples{
F2 <- qtl::read.cross(format="csv",
                      file = system.file("extdata", "F2maize_geno.csv",
                                        package = "RAP"),
                      genotypes = c("AA", "AB", "BB"),
                      alleles = c("A", "B"), estimate.map = FALSE)
## Perform a composite interval mapping for detecting QTLs.
QTLDet <- QTLDetect(cross = F2, trait = "trait", type = "CIM")
\dontrun{
## Create a pdf report summarizing the results.
report(QTLDet, outfile = "./testReports/reportQTLDectection.pdf")
}

}
