% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAMMI.R
\name{residuals.AMMI}
\alias{residuals.AMMI}
\title{Extract residuals.}
\usage{
\method{residuals}{AMMI}(object, ...)
}
\arguments{
\item{object}{An object of class AMMI}

\item{...}{Not used.}
}
\value{
A data.frame with residuals.
}
\description{
Extract the residuals for the fitted AMMI model.
}
\examples{
## Run AMMI analysis on TDMaize.
geAmmi <- gxeAmmi(TD = TDMaize, trait = "yld")

## Extract residuals.
residAmmi <- residuals(geAmmi)
head(residAmmi)

}
\seealso{
Other AMMI: 
\code{\link{fitted.AMMI}()},
\code{\link{gxeAmmi}()},
\code{\link{plot.AMMI}()},
\code{\link{report.AMMI}()}
}
\concept{AMMI}
