% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFW.R
\name{residuals.FW}
\alias{residuals.FW}
\title{Extract residuals.}
\usage{
\method{residuals}{FW}(object, ...)
}
\arguments{
\item{object}{An object of class FW}

\item{...}{Not used.}
}
\value{
A data.frame with residuals.
}
\description{
Extract the residuals for a fitted Finlay-Wilkinson model.
}
\examples{
## Run Finlay-Wilkinson analysis.
geFW <- gxeFw(TD = TDMaize, trait = "yld")

## Extract residuals.
residFW <- residuals(geFW)
head(residFW)

}
\seealso{
Other Finlay-Wilkinson: 
\code{\link{fitted.FW}()},
\code{\link{gxeFw}()},
\code{\link{plot.FW}()},
\code{\link{report.FW}()}
}
\concept{Finlay-Wilkinson}
