% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STExtract.R
\name{STExtract}
\alias{STExtract}
\title{Extract statistics from Fitted Models}
\usage{
STExtract(SSA, trials = names(SSA), traits = NULL, what = "all",
  keep = NULL)
}
\arguments{
\item{SSA}{An object of class SSA.}

\item{traits}{A character vector of traits for which the statistics should be
computed. If not supplied statistics are computed for all traits that have
been modelled.}

\item{what}{A character vector indicating which statistics should be
computed. Most statistics are available for all models, some only for models
fitted using a certain engine. If this is the case this is indicated in the
list with options in details.\cr
If \code{what = "all"} all available statistics are computed.}

\item{keep}{A character vector of column(s) in the object of class
\code{\link{TD}} used for modeling. These columns will be kept as output when
computing fitted values, residuals, standardized residuals and rMeans.
Columns can also be kept when computing (se)BLUEs and (se)BLUPs but only if
the column to keep contains unique values for the modeled variables, i.e. a
column repId with several different values per genotype cannot be kept.}
}
\value{
A list of extracted statistics or if only one statistic is extracted
a single object containing this statistic.
}
\description{
This function extracts and calculates various results for fitted models such
as BLUEs, BLUPs, unit errors and heritabilities. Note that most results can
only be calculated if a model is fitted with genotype as fixed or random.
This is indicated in the list below with "F" and "R"
}
\details{
Possible options for \code{what} are:
\describe{
\item{F - BLUEs}{Best Lineair Unbiased Estimators.}
\item{F - seBLUES}{Standard errors of the BLUEs.}
\item{R - BLUPs}{Best Lineair Unbiased Predictors.}
\item{R - seBLUPs}{Standard errors of the BLUPs.}
\item{F - ue}{Unit errors - only for \code{lme4} and \code{asreml}.}
\item{R - heritability}{Heritability.}
\item{R - varGen}{Genetic variance component.}
\item{R - varErr}{Residual variance component - only for \code{lme4}
and \code{asreml}.}
\item{R - varSpat}{Spatial variance components - only for \code{SpATS}.}
\item{F - fitted}{Fitted values for the model with genotype as fixed
component.}
\item{F - resid}{Residuals for the model with genotype as fixed component.}
\item{F - stdRes}{Standardized residuals for the model with genotype as fixed
component - only for \code{lme4} and \code{asreml}.}
\item{R - rMeans}{Fitted values for the model with genotype as random
component.}
\item{R - ranEf}{Random genetic effects.}
\item{F - wald}{Results of the wald test - only for \code{lme4} and
\code{asreml}.}
\item{F - CV}{Coefficient of variation - only for \code{lme4} and
\code{asreml}.}
\item{F - rDf}{Residual degrees of freedom.}
\item{R - effDim}{Effective dimensions - only for \code{SpATS}.}
\item{F - sed}{Standard error of difference - only for \code{asreml}.}
\item{F - lsd}{Least significant difference - only for \code{asreml}.}
\item{all}{All available statistics.}
}
}
\examples{
## Fit model using SpATS.
myModel <- STRunModel(TD = TDHeat05, design = "res.rowcol", traits = "yield")
## Extract all available statistics from the fitted model.
extr <- STExtract(myModel)
## Extract only the BLUEs from the fitted model.
BLUEs <- STExtract(myModel, what = "BLUEs")
## Extract only the BLUEs from the fitted model and keep trial as variable in
## the output.
BLUEs2 <- STExtract(myModel, what = "BLUEs", keep = "trial")

}
\seealso{
\code{\link{STRunModel}}, \code{\link{STModSpATS}}, \code{\link{STModLme4}}
and \code{\link{STModAsreml}}
}
