% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STModSommer.R
\name{STModSommer}
\alias{STModSommer}
\title{Fit Single Trial Model using sommer}
\usage{
STModSommer(TD, traits, what = c("fixed", "random"), covariates = NULL,
  useCheckId = FALSE, control = NULL, trySpatial = FALSE,
  design = "rowcol", ...)
}
\arguments{
\item{TD}{An object of class \code{\link{TD}}.}

\item{traits}{A character vector specifying the selected traits.}

\item{what}{A character vector specififying whether "genotype" should
be fitted as "fixed" or "random". If not specified both models
are fitted.}

\item{covariates}{A string specifying (a) covariate name(s).}

\item{useCheckId}{Should a checkId be used as a fixed parameter in the model?\cr
If \code{TRUE} \code{TD} has to contain a column 'checkId'.}

\item{control}{An optional list with control parameters to be passed to the
actual fitting funcions. For now only nSeg and nestDiv are valid and pass a
value to nseg and nest.div in \code{\link[SpATS]{PSANOVA}} respectively.}

\item{trySpatial}{Should spatial models be tried? Spatials models are can
only be fitted with SpATS and asreml.}

\item{design}{A string specifying the experimental design. One of "ibd"
(incomplete block design), "res.ibd" (resolvable incomplete block design),
"rcbd" (randomized complete block design), "rowcol" (row column design) or
"res.rowcol" (resolvable row column design).}

\item{...}{Further arguments to be passed to \code{SpATS}, \code{lme4} or
\code{asreml}.}
}
\description{
Fit Single Trial Model using sommer
}
\examples{
## Fit model for row column design.
STModSommer_1 <- STModSommer(TD = TDHeat05, trait = "yield")

## Fit model for row column including replicates.
STModSommer_2 <- STModSommer(TD = TDHeat05, trait = "yield",
                            design = "res.rowcol")

}
\seealso{
\code{\link{STRunModel}}
}
