% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSSA.R
\name{summary.SSA}
\alias{summary.SSA}
\title{Summarizing objects of class \code{SSA}}
\usage{
\method{summary}{SSA}(object, trial = NULL, trait = NULL,
  digits = max(getOption("digits") - 2, 3), nBest = 20, sortBy = if
  (!is.null(object$mFix)) "BLUEs" else "BLUPs", naLast = TRUE,
  decreasing = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{SSA}.}

\item{trait}{A character string indicating the trait to summarize. If
\code{trait = NULL} and only one trait is modelled this trait is summarized.}

\item{digits}{An integer indicating the number of significant digits for
printing.}

\item{nBest}{An integer indicating the number of the best genotypes (sorted
by either BLUEs or BLUPs) to print. If \code{NA} all genotypes will be
printed.}

\item{sortBy}{A character string specifying how the genotypes will be sorted.
Either \code{"BLUEs"}, \code{"BLUPs"} or \code{NA} (i.e. no sorting).}

\item{naLast}{Should missing values in the data be put last when sorting?}

\item{decreasing}{Should the sort order be decreasing?}

\item{...}{Further arguments passed to \code{\link[stats]{printCoefmat}}.}
}
\description{
\code{summary} method for class \code{SSA}.
}
\examples{
## Run a single trait analysis using SpATS.
myModel <- STRunModel(TD = TDHeat05, design = "res.rowcol", traits = "yield")
## Print a summary of the fitted model.
summary(myModel)

}
