% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTD.R
\name{summary.TD}
\alias{summary.TD}
\title{Summarizing objects of class \code{TD}}
\usage{
\method{summary}{TD}(object, ..., trial = names(object), traits,
  what = c("nObs", "nMiss", "mean", "median", "min", "max", "lowerQ",
  "upperQ", "var"))
}
\arguments{
\item{object}{An object of class TD.}

\item{...}{Further arguments - currently not used.}

\item{trial}{A character vector specifying the trials to be plotted.}

\item{traits}{A character vector specifying the traits to be summarised.}

\item{what}{A character vector indicating which summary statistics should be
computed. If \code{what = "all"} all available statistics are computed.\cr
Possible options are:
\describe{
\item{nVals}{The number of values, i.e. non-missing + missing values.}
\item{nObs}{The number of non-missing observations.}
\item{nMiss}{The number of missing values.}
\item{mean}{The mean.}
\item{median}{The median.}
\item{min}{The minimum.}
\item{max}{The maximum.}
\item{range}{The range (maximum - minimum).}
\item{lowerQ}{The lower (25\%) quantile.}
\item{upperQ}{The upper (75\%) quantile.}
\item{sd}{The standard deviation.}
\item{seMean}{The standard error of mean.}
\item{var}{The variance.}
\item{seVar}{The standard error of variance.}
\item{CV}{The coefficient of variation.}
\item{sum}{The sum.}
\item{sumSq}{The sum of squares.}
\item{uncorSumSq}{The uncorrected sum of squares.}
\item{skew}{The skewness.}
\item{seSkew}{The standard error of the skewness.}
\item{kurt}{The kurtosis.}
\item{seKurt}{The standard error of the kurtosis.}
\item{all}{All summary statistics.}
}}
}
\value{
A table containing the selected summary statistics.
}
\description{
\code{summary} method for class \code{TD}.
}
\examples{
## Summarize TDHeat05.
summary(object = TDHeat05, traits = "yield")

}
\seealso{
\code{\link{createTD}}
}
