% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTD.R
\name{TD}
\alias{TD}
\alias{createTD}
\title{S3 class TD}
\usage{
createTD(data, genotype = NULL, env = NULL, megaEnv = NULL, year = NULL,
  repId = NULL, subBlock = NULL, rowId = NULL, colId = NULL,
  rowCoordinates = NULL, colCoordinates = NULL, checkId = NULL,
  design = NULL)
}
\arguments{
\item{data}{a data.frame containing trial data with a least a column for
genotype.}

\item{genotype}{an optional character string indicating the column in
\code{data} that contains genotypes.}

\item{env}{an optional character string indicating the column in \code{data}
that contains environments.}

\item{megaEnv}{an optional character string indicating the column in
\code{data} that contains megaEnvironments as constructed by
\code{\link{gxeMegaEnv}}.}

\item{year}{an optional character string indicating the column in \code{data}
that contains years.}

\item{repId}{an optional character string indicating the column in \code{data}
that contains replicates.}

\item{subBlock}{an optional character string indicating the column in
\code{data} that contains sub blocks.}

\item{rowId}{an optional character string indicating the column in
\code{data} that contains field rows.}

\item{colId}{an optional character string indicating the column in
\code{data} that contains field columns.}

\item{rowCoordinates}{an optional character string indicating the column in
\code{data} that contains the rowId coordinates used for fitting spatial
models.}

\item{colCoordinates}{an optional character string indicating the column in
\code{data} that contains the column coordinates used for fitting spatial
models.}

\item{checkId}{an optional character string indicating the column in
\code{data} that contains the check ID(s).}

\item{design}{an optional character string indicating the design of the
trial. Accepted values are "ibd" (incomplete-block design), "res.ibd"
(resolvable incomplete-block design), "rcbd" (randomized complete block
design), "rowcol" (row-column design) and "res.rowcol" (resolvable
row-column design).}
}
\value{
An object of class TD, the input data.frame with renamed columns
and an attribute \code{renamedCols} containing info on which columns have
been renamed.
}
\description{
Function for creating objects of S3 class TD (Trial Data). The input data is
checked and columns are renamed to default column names for ease of further
computations. The columns for genotype, env, megaEnv, year, repId, subBlock,
rowId, colId and checkId are converted to factor columns, whereas
rowCoordinates and colCoordinates are converted to numerical columns. One
single column can be mapped to multiple defaults, e.g. one column with x
coordinates can be mapped to both colId and colCoordinates.\cr
Columns other than the default columns, e.g. traits or other covariates
will be included in the output unchanged.
}
\details{
\code{\link{print}} and \code{\link{summary}} and \code{\link{plot}} methods
are available.
}
\seealso{
\code{\link{summary.TD}}, \code{\link{plot.TD}}
}
\author{
Bart-Jan van Rossum
}
