% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTD.R
\name{TD}
\alias{TD}
\alias{createTD}
\alias{addTD}
\alias{dropTD}
\title{S3 class TD}
\usage{
createTD(data, genotype = NULL, trial = NULL, megaEnv = NULL,
  year = NULL, repId = NULL, subBlock = NULL, rowId = NULL,
  colId = NULL, rowCoordinates = NULL, colCoordinates = NULL,
  checkId = NULL, design = NULL)

addTD(TD, data, genotype = NULL, trial = NULL, megaEnv = NULL,
  year = NULL, repId = NULL, subBlock = NULL, rowId = NULL,
  colId = NULL, rowCoordinates = NULL, colCoordinates = NULL,
  checkId = NULL, design = NULL)

dropTD(TD, trials)
}
\arguments{
\item{data}{A data.frame containing trial data with a least a column for
genotype.}

\item{genotype}{An optional character string indicating the column in
\code{data} that contains genotypes.}

\item{trial}{An optional character string indicating the column in \code{data}
that contains trials.}

\item{megaEnv}{An optional character string indicating the column in
\code{data} that contains mega-environments as constructed by
\code{\link{gxeMegaEnv}}.}

\item{year}{An optional character string indicating the column in \code{data}
that contains years.}

\item{repId}{An optional character string indicating the column in \code{data}
that contains replicates.}

\item{subBlock}{An optional character string indicating the column in
\code{data} that contains sub blocks.}

\item{rowId}{An optional character string indicating the column in
\code{data} that contains field rows.}

\item{colId}{An optional character string indicating the column in
\code{data} that contains field columns.}

\item{rowCoordinates}{An optional character string indicating the column in
\code{data} that contains the rowId coordinates used for fitting spatial
models.}

\item{colCoordinates}{An optional character string indicating the column in
\code{data} that contains the column coordinates used for fitting spatial
models.}

\item{checkId}{An optional character string indicating the column in
\code{data} that contains the check IDs.}

\item{design}{An optional character string indicating the design of the
trial. Either are "ibd" (incomplete-block design), "res.ibd"
(resolvable incomplete-block design), "rcbd" (randomized complete block
design), "rowcol" (row-column design) or "res.rowcol" (resolvable
row-column design).}

\item{TD}{An object of class TD to which the data should be added.}

\item{trials}{a character vector of trials that should be removed.}
}
\value{
An object of class TD, the input data.frame with renamed columns
and an attribute \code{renamedCols} containing info on which columns have
been renamed.
}
\description{
Function for creating objects of S3 class TD (Trial Data). The input data is
checked and columns are renamed to default column names for ease of further
computations. The columns for genotype, trial, megaEnv, year, repId,
subBlock, rowId, colId and checkId are converted to factor columns, whereas
rowCoordinates and colCoordinates are converted to numerical columns. One
single column can be mapped to multiple defaults, e.g. one column with x
coordinates can be mapped to both colId and colCoordinates.\cr
Columns other than the default columns, e.g. traits or other covariates
will be included in the output unchanged.\cr
\code{\link{print}} and \code{\link{summary}} and \code{\link{plot}} methods
are available.

Function for adding extra data to an existing object of class TD.

Function for removing data for selected trials from an existing object of
class TD.
}
\seealso{
\code{\link{summary.TD}}, \code{\link{plot.TD}}
}
\author{
Bart-Jan van Rossum
}
