% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDHeat05.R
\docType{data}
\name{TDHeat05}
\alias{TDHeat05}
\title{Test data}
\format{An object of class \code{TD} (inherits from \code{data.frame}) with 338 rows and 10 columns.}
\usage{
TDHeat05
}
\description{
Test data
}
\keyword{datasets}
