% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDHeat05.R
\docType{data}
\name{testBc}
\alias{testBc}
\title{Fake data of backcross for unit testing.}
\format{An object of class \code{bc} (inherits from \code{cross}) of length 2.}
\usage{
testBc
}
\description{
Fake data of backcross for unit testing.
}
\keyword{datasets}
