% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarComp.R
\name{varComp}
\alias{varComp}
\alias{createVarComp}
\title{S3 class varComp}
\usage{
createVarComp(STA, choice, summary, vcov, criterion, engine)
}
\arguments{
\item{STA}{An object of class STA, the best fitted model.}

\item{choice}{A character string indicating the best fitted model.}

\item{summary}{A data.frame with a summary of the fitted models.}

\item{vcov}{The covariance matrix of the best fitted model.}

\item{criterion}{A character string indicating the goodness-of-fit criterion
used for determinening the best model.}

\item{engine}{A character string containing the engine used for the analysis.}
}
\description{
Function for creating objects of S3 class varComp.\cr
\code{\link{print}}, \code{\link{summary}}, \code{\link{plot}} and
\code{\link{report}} methods are available.
}
\seealso{
\code{\link{plot.varComp}}, \code{\link{report.varComp}}
}
\keyword{internal}
